% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata_to_daisie_data.R
\name{add_metadata_to_daisie_data}
\alias{add_metadata_to_daisie_data}
\title{Formats simulation output into standard DAISIE list output.}
\usage{
add_metadata_to_daisie_data(multi_daisie_data, total_time, m)
}
\arguments{
\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{m}{Numeric defining the size of mainland pool.}
}
\value{
List with DAISIE simulation output
}
\description{
Formats simulation output into standard DAISIE list output.
}
\author{
Joshua W. Lambert
}
\keyword{internal}
