% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_colonisations_to_table.R
\name{all_colonisations_to_table}
\alias{all_colonisations_to_table}
\title{Extract all colonisations from an island history
and put these in a table}
\usage{
all_colonisations_to_table(ideal_or_empirical_island)
}
\arguments{
\item{ideal_or_empirical_island}{the evolutionary history of
an island from ideal or empirical data,
as created by \link{sim_island}

\code{ideal_or_empirical_island} is a list  containing 3 components:
\itemize{
\item \verb{$branching_times}: island age, species stem age and speciation times.
See \link{check_branching_times} for more details
\item \verb{$stac}: the status of the colonist,
see \link{stac_to_str} for values and meanings
\item \verb{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)
}

For recolonising lineages, there is an extra element,
\code{all_colonisations} per list element.
It is comprised of \verb{$event_times} and \verb{$species_type}:
\describe{
\item{\verb{$event_times}}{ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times. Each vector pertains to one
colonising lineage.}
\item{\verb{$species_type}}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}}
}
\value{
a table with column names:
\itemize{
\item \code{clade_id}: the clade ID
\item \code{colonist_id}: the colonists' ID
\item \code{event_times} ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times
\item \code{species_type} a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.
}
}
\description{
Extract all colonisations from an island history
and put these in a table
}
\examples{
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = 20)
island <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 12, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
all_colonisations_to_table(ideal_or_empirical_island = island$ideal_island)
}
\author{
Richèl J.C. Bilderbeek
}
