% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_endemic_clades.R
\name{all_endemic_clades}
\alias{all_endemic_clades}
\title{Checks if all island species are endemic clades. Clade are colonisations that
have formed an island clade of two or more species surviving to the present.}
\usage{
all_endemic_clades(island)
}
\arguments{
\item{island}{The history of a a single island,
as can be checked by \link{check_island}.

An \code{island} is \link{list} with two elements:
\itemize{
\item \code{ideal_island}: the island history based on ideal data,
as can be checked by \link{check_ideal_island}
\item \code{empirical_island}: the island history based on empirical data
as can be checked by \link{check_empirical_island}
For recolonising lineages, there is an extra element,
\code{all_colonisations} per list element.
It is comprised of \verb{$event_times} and \verb{$species_type}:
\describe{
\item{\verb{$event_times}}{ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times. Each vector pertains to one
colonising lineage.}
\item{\verb{$species_type}}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}
}}
}
\value{
Boolean
}
\description{
Checks if all island species are endemic clades. Clade are colonisations that
have formed an island clade of two or more species surviving to the present.
}
\examples{
\dontrun{
island <- DAISIEmainland::sim_island_with_mainland(
  total_time = 1,
  m = 100,
  island_pars = c(1,1,10,0.1,1),
  mainland_ex = 0.1,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete",
  replicates = 2,
  verbose = FALSE)
bool <- all_endemic_clades(island$ideal_islands[[1]])
}
}
\author{
Joshua W. Lambert
}
