% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ana_event.R
\name{ana_event}
\alias{ana_event}
\title{Updates the island_spec and max_spec_id when an anagenesis event occurs}
\usage{
ana_event(island_spec, max_spec_id)
}
\arguments{
\item{island_spec}{Matrix with current state of simulation containing number
of species.}

\item{max_spec_id}{Current species IDs.}
}
\value{
A two element list
}
\description{
Updates the island_spec and max_spec_id when an anagenesis event occurs
}
\author{
Joshua W. Lambert
}
