% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_recols.R
\name{any_recols}
\alias{any_recols}
\title{Determines if the island contains any recolonisations that survived to the
present}
\usage{
any_recols(island)
}
\arguments{
\item{island}{The history of a a single island,
as can be checked by \link{check_island}.

An \code{island} is \link{list} with two elements:
\itemize{
\item \code{ideal_island}: the island history based on ideal data,
as can be checked by \link{check_ideal_island}
\item \code{empirical_island}: the island history based on empirical data
as can be checked by \link{check_empirical_island}
}}
}
\value{
Boolean
}
\description{
Determines if the island contains any recolonisations that survived to the
present
}
\examples{
\dontrun{
island <- DAISIEmainland::sim_island_with_mainland(
  total_time = 1,
  m = 100,
  island_pars = c(1,1,10,0.1,1),
  mainland_ex = 0.1,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete",
  replicates = 2,
  verbose = FALSE)
bool <- any_recols(island$ideal_islands[[1]])
}
}
