% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_code_to_y.R
\name{branch_code_to_y}
\alias{branch_code_to_y}
\title{Convert a \code{branch_code} to a y coordinat in range \link{0, 1}}
\usage{
branch_code_to_y(branch_code)
}
\arguments{
\item{branch_code}{a species' branching pattern,
as checked by \link{check_branch_code}, e.g. \code{A}, \code{AB}, \code{AABA}.}
}
\value{
a y coordinat in range \link{0, 1}
}
\description{
A valid \code{branch_code} ...
}
\details{
\itemize{
\item is a string, i.e. a character vector with one element
\item consists out of characters that are either \code{A} or \code{B}
\item starts with \code{A}
}
}
\examples{
branch_code_to_y("A")
branch_code_to_y("AA")
branch_code_to_y("AB")
}
\author{
Richèl J.C. Bilderbeek
}
