% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_empirical_col.R
\name{calc_empirical_col}
\alias{calc_empirical_col}
\title{Calculates whether colonisations from different mainland species are more}
\usage{
calc_empirical_col(island_spec, mainland_clade)
}
\arguments{
\item{island_spec}{Matrix with current state of simulation containing number
of species, see \code{\link[=create_test_island_spec]{create_test_island_spec()}}}

\item{mainland_clade}{Data frame with state of mainland, see
\code{\link[=create_test_mainland_clade]{create_test_mainland_clade()}}}
}
\value{
A list of numerics or numeric vector
}
\description{
Calculates whether colonisations from different mainland species are more
}
