% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_endemic_ratio.R
\name{calc_endemic_ratio}
\alias{calc_endemic_ratio}
\title{Title}
\usage{
calc_endemic_ratio(daisie_data)
}
\arguments{
\item{daisie_data}{stub}
}
\value{

}
\description{
Title
}
\author{
Joshua W. Lambert
}
