% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error.R
\name{calc_error}
\alias{calc_error}
\title{Calculates the ...}
\usage{
calc_error(daisie_data, ideal_ml, empirical_ml)
}
\arguments{
\item{daisie_data}{stub}

\item{ideal_ml}{stub}

\item{empirical_ml}{stub}
}
\value{

}
\description{
Calculates the ...
}
\author{
Joshua W. Lambert
}
