% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_island_endemics.R
\name{calc_island_endemics}
\alias{calc_island_endemics}
\title{Calculates the number of endemic and non-endemics species on a single island
replicate}
\usage{
calc_island_endemics(daisie_data)
}
\arguments{
\item{daisie_data}{\link{list} containing data of of a DAISIE simulation
with mainland dynamics,
as produced by \link{sim_island_with_mainland}
and checked by \link{check_daisie_data}

A \code{daisie_data} has two elements:
\itemize{
\item \code{ideal_islands}: the ideal data set produced in a simulation,
see \link{check_ideal_daisie_data} for more details
\item \code{empirical_islands}: the empirical data sets produced in a simulation
see \link{check_empirical_daisie_data} for more details
}}
}
\value{
List of two numerics
}
\description{
Calculates the number of endemic and non-endemics species on a single island
replicate
}
\author{
Joshua W. Lambert
}
