% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_max_age_percent.R
\name{calc_max_age_percent}
\alias{calc_max_age_percent}
\title{Calculates the percent of max age species on the island for each replicate}
\usage{
calc_max_age_percent(daisie_data)
}
\arguments{
\item{daisie_data}{List containing data of DAISIE simulation with mainland
dynamics. Output from `sim_island_with_mainland``}
}
\value{
A list of two numeric vectors
}
\description{
Calculates the percent of max age species on the island for each replicate
}
\author{
Joshua W. Lambert
}
