% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_max_age_ratio.R
\name{calc_max_age_ratio}
\alias{calc_max_age_ratio}
\title{Title}
\usage{
calc_max_age_ratio(daisie_data)
}
\arguments{
\item{daisie_data}{stub}
}
\value{

}
\description{
Title
}
\author{
Joshua W. Lambert
}
