% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{calc_outliers}
\alias{calc_outliers}
\title{Subsets the data to those less than the 5th percentile and those greater than
the 95th percentile}
\usage{
calc_outliers(plotting_data)
}
\arguments{
\item{plotting_data}{Numeric vector of data points}
}
\value{
Numeric vector
}
\description{
Subsets the data to those less than the 5th percentile and those greater than
the 95th percentile
}
