% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_overall_sim_metrics.R
\name{calc_overall_sim_metrics}
\alias{calc_overall_sim_metrics}
\title{Calculates the summary metrics for the entire parameter space}
\usage{
calc_overall_sim_metrics(data_folder_path, output_file_path, param_space)
}
\arguments{
\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{param_space}{String specifying which parameter space results to read.
Either "general", "mainland_ex", or "mainland_sample_prob"}
}
\value{
List of simulation metrics
}
\description{
Calculates the summary metrics for the entire parameter space
}
\author{
Joshua W. Lambert
}
