% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_param_diffs.R
\name{calc_param_diffs}
\alias{calc_param_diffs}
\title{Title}
\usage{
calc_param_diffs(ideal_ml, empirical_ml)
}
\arguments{
\item{ideal_ml}{stub}

\item{empirical_ml}{stub}
}
\value{

}
\description{
Title
}
\author{
Joshua W. Lambert
}
