% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{calc_quantiles}
\alias{calc_quantiles}
\title{Calculates the 5th, 25th, 50th, 75th and 95th percentile of the data}
\usage{
calc_quantiles(plotting_data)
}
\arguments{
\item{plotting_data}{Numeric vector of data points}
}
\value{
Named vector of 5 named elements
}
\description{
Calculates the 5th, 25th, 50th, 75th and 95th percentile of the data
}
