% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all_colonisations.R
\name{check_all_colonisations}
\alias{check_all_colonisations}
\title{Check if a \code{all_colonisations} is valid.}
\usage{
check_all_colonisations(all_colonisations)
}
\value{
nothing
}
\description{
Check if a \code{all_colonisations} is valid.
Will \link{stop} if not
}
\details{
\code{all_colonisations} contains:
\itemize{
\item for cladogenetic species: the island age
and branching times of the radiation,
including the stem age of the radiation
\item for non-endemic, non-endemic_MaxAge and
Endemic anagenetic species:
island age and stem age of the population/species
}

A valid \code{all_colonisations} ...
\itemize{
\item is numeric
\item has only finite values
\item has at least 1 element (which equals the island age)
\item has only positive non-zero values (although zero is accepted, to prevent
false negatives by rounding off errors)
\item has a descending order
}
}
\examples{
check_all_colonisations(1.0)
check_all_colonisations(c(1.0, 0.25))
}
\author{
Richèl J.C. Bilderbeek
}
