% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_analysis_result.R
\name{check_analysis_result}
\alias{check_analysis_result}
\title{Checks whet}
\usage{
check_analysis_result(analysis_results)
}
\arguments{
\item{analysis_results}{}
}
\value{

}
\description{
Checks whet
}
