% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_branch_code.R
\name{check_branch_code}
\alias{check_branch_code}
\title{Check if a \code{branch_code} is valid.
Will \link{stop} if not}
\usage{
check_branch_code(branch_code)
}
\arguments{
\item{branch_code}{a species' branching pattern,
as checked by \link{check_branch_code}, e.g. \code{A}, \code{AB}, \code{AABA}.}
}
\value{
nothing
}
\description{
Check if a \code{branch_code} is valid.
Will \link{stop} if not
}
\examples{
check_branch_code("A")
check_branch_code("AA")
check_branch_code("AB")
}
\author{
Richèl J.C. Bilderbeek
}
