% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_branching_times.R
\name{check_branching_times}
\alias{check_branching_times}
\title{Check if a \code{branching_times} is valid.}
\usage{
check_branching_times(branching_times)
}
\arguments{
\item{branching_times}{island age and stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge and
Endemic anagenetic species.

For cladogenetic species these should
be island age and branching times of the radiation including the
stem age of the radiation.

See \link{check_branching_times} for a more detailed description.}
}
\value{
nothing
}
\description{
Check if a \code{branching_times} is valid.
Will \link{stop} if not
}
\details{
\code{branching_times} contains:
\itemize{
\item for cladogenetic species: the island age
and branching times of the radiation,
including the stem age of the radiation
\item for non-endemic, non-endemic_MaxAge and
Endemic anagenetic species:
island age and stem age of the population/species
}

A valid \code{branching_times} ...
\itemize{
\item is numeric
\item has only finite values
\item has at least 1 element (which equals the island age)
\item has only positive non-zero values (although zero is accepted, to prevent
false negatives by rounding off errors)
\item has a descending order
}
}
\examples{
check_branching_times(1.0)
check_branching_times(c(1.0, 0.25))
}
\author{
Richèl J.C. Bilderbeek
}
