% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_colonisations.R
\name{check_colonisations}
\alias{check_colonisations}
\title{Check if an \code{colonisations} is valid.}
\usage{
check_colonisations(colonisations)
}
\arguments{
\item{colonisations}{a \code{colonisations}, which is a \link{list} with elements:
\itemize{
\item \code{event_times}: the times that events (i.e. island age, immigration,
speciation) occurred (see \link{check_event_times})
\item \code{species_type}:  the origin of a clade (i.e. anagenetic, cladogenetic
or immigrant) (see \link{check_species_type})
}

An \code{all_colonisations} (see \link{check_all_colonisations}) is a \link{list}
of \code{colonisations}.}
}
\value{
nothing
}
\description{
Check if an \code{colonisations} is valid.
Will \link{stop} if not.
}
\details{
A \code{colonisations} contains:
\itemize{
\item a \code{species_type}, which is the origin of the clade,
which can be anagenetic, cladogenetic or immigration.
See \link{check_species_type} for details
\item a \code{event_times}, which holds events such as the island age,
stem age of the species and -for cladogenetic species only-,
the branching times of the radiation.
see \link{check_event_times}
}
}
\examples{
set.seed(
  2,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = 20
)
island_tbl <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 12, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
daisie_data <- DAISIEmainland:::create_daisie_data(
  total_time = 1,
  island_tbl = island_tbl,
  mainland_clade = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete"
)
colonisations <- daisie_data$ideal_island[[2]]$all_colonisations[[1]]
check_colonisations(colonisations)
}
\author{
Richèl J.C. Bilderbeek
}
