% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_daisie_data.R
\name{check_daisie_data}
\alias{check_daisie_data}
\title{Check if a \code{daisie_data} is valid.
Will \link{stop} if not.}
\usage{
check_daisie_data(daisie_data)
}
\arguments{
\item{daisie_data}{\link{list} containing data of of a DAISIE simulation
with mainland dynamics,
as produced by \link{sim_island_with_mainland}
and checked by \link{check_daisie_data}

A \code{daisie_data} has two elements:
\itemize{
\item \code{ideal_islands}: the ideal data set produced in a simulation,
see \link{check_ideal_islands} for more details
\item \code{empirical_islands}: the empirical data sets produced in a simulation
see \link{check_empirical_islands} for more details
}}
}
\value{
nothing
}
\description{
Check if a \code{daisie_data} is valid.
Will \link{stop} if not.
}
\examples{
daisie_data <- sim_island_with_mainland(
  total_time = 1,
  m = 100,
  island_pars = c(1, 1, 50, 0.1, 1),
  mainland_ex = 0.5,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete",
  replicates = 1,
  verbose = FALSE
)
check_daisie_data(daisie_data)

}
\author{
Richèl J.C. Bilderbeek
}
