% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_daisie_mainland_data.R
\name{check_daisie_mainland_data}
\alias{check_daisie_mainland_data}
\title{Check if a \code{daisie_mainland_data} is valid.
Will \link{stop} if not.}
\usage{
check_daisie_mainland_data(daisie_mainland_data)
}
\value{
nothing
}
\description{
Check if a \code{daisie_mainland_data} is valid.
Will \link{stop} if not.
}
\examples{
daisie_mainland_data <- sim_island_with_mainland(
  total_time = 1,
  m = 100,
  island_pars = c(1, 1, 50, 0.1, 1),
  mainland_ex = 0.5,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete",
  replicates = 1,
  verbose = FALSE
)
check_daisie_mainland_data(daisie_mainland_data)

}
\author{
Richèl J.C. Bilderbeek, Joshua W. Lambert
}
