% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_empirical_island.R
\name{check_empirical_island}
\alias{check_empirical_island}
\title{Check if an \code{empirical_island} is valid.}
\usage{
check_empirical_island(empirical_island)
}
\arguments{
\item{empirical_island}{a simulated island with imperfect information,
with one \code{empirical_island_clade} per simulated island clade.

An \code{empirical_island} is created by \link{sim_island}
(together with \code{ideal_island}, which has perfect information),
as can be checked by \link{check_empirical_island}.}
}
\value{
nothing
}
\description{
Check if an \code{empirical_island} is valid.
Will \link{stop} if not.
}
\examples{
set.seed(
  2,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = 2)
island <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 1, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")

check_empirical_island(island$empirical_island)
}
\seealso{
Use \link{check_ideal_island} to check if an \code{ideal_island}
is valid.
}
\author{
Richèl J.C. Bilderbeek
}
