% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_event_times.R
\name{check_event_times}
\alias{check_event_times}
\title{Check if a \code{event_times} is valid.}
\usage{
check_event_times(event_times)
}
\arguments{
\item{event_times}{the island age, stem age of species,
and -for cladogenetic species- speciation times,
as checked by \link{check_event_times}}
}
\value{
nothing
}
\description{
Check if a \code{event_times} is valid.
Will \link{stop} if not
}
\details{
\code{event_times} contains:
\itemize{
\item for cladogenetic species: the island age
and event times of the radiation,
including the stem age of the radiation
\item for non-endemic, non-endemic_MaxAge and
Endemic anagenetic species:
island age and stem age of the population/species
}

A valid \code{event_times} ...
\itemize{
\item is numeric
\item has only finite values
\item has at least 1 element (which equals the island age)
\item has only positive non-zero values (although zero is accepted, to prevent
false negatives by rounding off errors)
\item has a descending order
}
}
\examples{
check_event_times(1.0)
check_event_times(c(1.0, 0.25))
}
\author{
Richèl J.C. Bilderbeek
}
