% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ideal_daisie_data.R
\name{check_ideal_daisie_data}
\alias{check_ideal_daisie_data}
\title{Check if an \code{ideal_daisie_data} is valid.
Will \link{stop} if not.}
\usage{
check_ideal_daisie_data(ideal_daisie_data)
}
\arguments{
\item{ideal_daisie_data}{a \link{list},
of which each element contains the history
of a DAISIE simulation with mainland dynamics, with a length that
equals the number of replicates,
as can be checked by \link{check_ideal_daisie_data}

The simulated history is recorded perfectly, resulting in ideal data.
The list elements are of type \code{DAISIE::datalist},
as checked by \link{check_daisie_datalist}
and can be used by \link[DAISIE]{DAISIE} (e.g. \link[DAISIE]{DAISIE_ML_CS})}
}
\value{
nothing
}
\description{
Check if an \code{ideal_daisie_data} is valid.
Will \link{stop} if not.
}
\examples{
daisie_data <- sim_island_with_mainland(
  total_time = 1,
  m = 100,
  island_pars = c(1, 1, 50, 0.1, 1),
  mainland_ex = 0.5,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete",
  replicates = 1,
  verbose = FALSE
)
check_ideal_daisie_data(daisie_data$ideal_islands)

}
\author{
Richèl J.C. Bilderbeek
}
