% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ideal_island.R
\name{check_ideal_island}
\alias{check_ideal_island}
\title{Check if an \code{ideal_island} is valid.}
\usage{
check_ideal_island(ideal_island)
}
\arguments{
\item{ideal_island}{a simulated island with perfect information,
which is a \link{list} of one or more \code{ideal_island_clade}s.

An \code{ideal_island} is created by \link{sim_island} (together with \code{empirical_island},
which has imperfect information),
as can be checked by \link{check_ideal_island}}
}
\value{
nothing
}
\description{
Check if an \code{ideal_island} is valid.
Will \link{stop} if not.
}
\examples{
set.seed(
  2,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = 2)
island <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 1, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")

check_ideal_island(island$ideal_island)
}
\seealso{
Use \link{check_empirical_island} to check if an \code{empirical_island}
is valid.
}
\author{
Richèl J.C. Bilderbeek
}
