% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ideal_island_clade.R
\name{check_ideal_island_clade}
\alias{check_ideal_island_clade}
\title{Check if an \code{ideal_island_clade} is valid.}
\usage{
check_ideal_island_clade(ideal_island_clade)
}
\arguments{
\item{ideal_island_clade}{A simulated island clade with perfect information,
which is a \link{list} of one or more \code{ideal_island_clade}s,
as can be checked by \link{check_ideal_island_clade}

\code{ideal_island_clade} is a list  containing 3 components:
\itemize{
\item \verb{$branching_times}: island age, species stem age and speciation times.
See \link{check_branching_times} for more details
\item \verb{$stac}: the status of the colonist,
see \link{stac_to_str} for values and meanings
\item \verb{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)
}

For recolonising lineages, there is an extra element,
\code{all_colonisations} per list element.
It is comprised of \verb{$event_times} and \verb{$species_type}:
\describe{
\item{\verb{$event_times}}{ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times. Each vector pertains to one
colonising lineage.}
\item{\verb{$species_type}}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}}
}
\value{
nothing
}
\description{
Check if an \code{ideal_island_clade} is valid.
Will \link{stop} if not.
}
\examples{
set.seed(
  2,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = 2
)
island_tbl <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 1, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete"
)
daisie_data <- DAISIEmainland:::create_daisie_data(
  total_time = 1,
  island_tbl = island_tbl,
  mainland_clade = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete"
)

ideal_island <- daisie_data$ideal_island
#  First clade
ideal_island_clade <- ideal_island[[1]]
check_ideal_island_clade(ideal_island_clade)
}
\seealso{
Use \link{check_empirical_island_clade} to check
if an \code{empirical_island_clade} is valid.
}
\author{
Richèl J.C. Bilderbeek
}
