% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_island_replicates.R
\name{check_island_replicates}
\alias{check_island_replicates}
\title{Check if an \code{island_replicates} is valid.}
\usage{
check_island_replicates(island_replicates)
}
\arguments{
\item{island_replicates}{a \link{list} of that has as many elements as
replicates. Each element is a \link{list} of \code{islands} (as checked
by \link{check_island}).
Use \link{check_island_replicates} to check a \code{island_replicates}
being correct}
}
\value{
nothing
}
\description{
Check if an \code{island_replicates} is valid.
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
