% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_island_state.R
\name{check_island_state}
\alias{check_island_state}
\title{Checks if any changes on the mainland change the state of the island species.}
\usage{
check_island_state(timeval, totaltime, island_spec, mainland)
}
\arguments{
\item{timeval}{Numeric defining current time of simulation.}

\item{totaltime}{Numeric defining the length of the simulation in time units.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}

\item{mainland}{Matrix with state of mainland.}
}
\value{
List
}
\description{
Checks if any changes on the mainland change the state of the island species.
}
