% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_island_tbl.R
\name{check_island_tbl}
\alias{check_island_tbl}
\title{Check if an \code{island_tbl} is valid.}
\usage{
check_island_tbl(island_tbl)
}
\arguments{
\item{island_tbl}{Data frame with current state of simulation containing
number of species, see \code{\link[=create_test_island_tbl]{create_test_island_tbl()}}}
}
\value{
nothing
}
\description{
Check if an \code{island_tbl} is valid.
Will \link{stop} if not.
}
\examples{
set.seed(
  2,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = 2)
island_tbl <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 1, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")

check_island(island_tbl)
}
\author{
Richèl J.C. Bilderbeek, Joshua W. Lambert
}
