% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mainland_clade.R
\name{check_mainland_clade}
\alias{check_mainland_clade}
\title{Check if an \code{mainland_clade} is valid.}
\usage{
check_mainland_clade(mainland_clade)
}
\arguments{
\item{mainland_clade}{Data frame with state of mainland, see
\code{\link[=create_test_mainland_clade]{create_test_mainland_clade()}}}
}
\value{
nothing
}
\description{
Check if an \code{mainland_clade} is valid.
Will \link{stop} if not.
}
\examples{
set.seed(
  2,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)

multi_mainland_clade <- DAISIEmainland::sim_mainland(
  total_time = 1,
  m = 2,
  mainland_ex = 1)

check_mainland_clade(mainland_clade)
}
\author{
Joshua W. Lambert, Richèl J.C. Bilderbeek
}
