% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_species.R
\name{check_missing_species}
\alias{check_missing_species}
\title{Check if a \code{missing_species} is valid.}
\usage{
check_missing_species(missing_species)
}
\arguments{
\item{missing_species}{the number of island species that were not
sampled for particular clade (only applicable for endemic clades),
as can be checked by \link{check_missing_species}}
}
\value{
nothing
}
\description{
Check if a \code{missing_species} is valid.
Will \link{stop} if not
}
\details{
A valid \code{missing_species} ...
\itemize{
\item is numeric
\item has one element
\item the element has a whole-number value in range \verb{[0, Inf>}
}
}
\examples{
check_missing_species(1)
}
\author{
Richèl J.C. Bilderbeek
}
