% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_multi_daisie_data.R
\name{check_multi_daisie_data}
\alias{check_multi_daisie_data}
\title{Check if an \code{multi_daisie_data} is valid for both \code{ideal_multi_daisie_data}
and \code{empirical_multi_daisie_data}.
Will \link{stop} if not.}
\usage{
check_multi_daisie_data(multi_daisie_data)
}
\arguments{
\item{multi_daisie_data}{A list of \code{daisie_data} elements}
}
\value{
nothing
}
\description{
An \code{multi_daisie_data} ...
}
\details{
\itemize{
\item is a \link{list} of \code{daisie_data}, see \link{check_daisie_data}
\item has at least 1 element
}
}
\examples{
daisie_mainland_data <- sim_island_with_mainland(
  total_time = 1,
  m = 100,
  island_pars = c(1, 1, 50, 0.1, 1),
  mainland_ex = 0.5,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete",
  replicates = 1,
  verbose = FALSE
)
check_multi_daisie_data(daisie_mainland_data$ideal_multi_daisie_data)
check_multi_daisie_data(daisie_mainland_data$empirical_multi_daisie_data)

}
\author{
Richèl J.C. Bilderbeek, Joshua W. Lambert
}
