% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_multi_mainland_clade.R
\name{check_multi_mainland_clade}
\alias{check_multi_mainland_clade}
\title{Check if an \code{multi_mainland_clade} is valid.
Will \link{stop} if not.}
\usage{
check_multi_mainland_clade(multi_mainland_clade)
}
\arguments{
\item{multi_mainland_clade}{A list of data frames each data frame is an
\code{mainland_clade}}
}
\value{
nothing
}
\description{
An \code{multi_mainland_clade} ...
}
\details{
\itemize{
\item is a \link{list} of \code{mainland_clade}, see \link{check_mainland_clade}
\item has at least 1 element
}
}
\examples{
multi_mainland_clade <- sim_mainland(
  total_time = 1,
  m = 10,
  mainland_ex = 0.5
)
check_multi_mainland_clade(multi_mainland_clade)

}
\author{
Richèl J.C. Bilderbeek, Joshua W. Lambert
}
