% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_species_type.R
\name{check_species_type}
\alias{check_species_type}
\title{Check if a \code{species_type} is valid.}
\usage{
check_species_type(species_type)
}
\arguments{
\item{species_type}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}
\value{
nothing
}
\description{
Check if a \code{species_type} is valid.
Will \link{stop} if not
}
\details{
A valid \code{species_type} ...
\itemize{
\item is numeric
\item has one element
\item the element has a whole-number value in range \verb{[1, 6]}
}
}
\examples{
check_species_type("A")
check_species_type("C")
check_species_type("I")
}
\seealso{
Use \link{species_type_to_str} to convert a \code{species_type} to a string.
}
\author{
Richèl J.C. Bilderbeek
}
