% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stac.R
\name{check_stac}
\alias{check_stac}
\title{Check if a \code{stac} is valid.}
\usage{
check_stac(stac)
}
\arguments{
\item{stac}{the status of the colonist,
see \link{stac_to_str} for values and meanings}
}
\value{
nothing
}
\description{
Check if a \code{stac} is valid.
Will \link{stop} if not
}
\details{
A valid \code{stac} ...
\itemize{
\item is numeric
\item has one element
\item the element has a whole-number value in range \verb{[0, 6]}
}
}
\examples{
check_stac(0)
check_stac(1)
check_stac(6)
}
\seealso{
Use \link{stac_to_str} to convert a \code{stac} to a string.
}
\author{
Richèl J.C. Bilderbeek
}
