% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{choose_scientific}
\alias{choose_scientific}
\title{Decides whether number should be in normal or scientific form depending on
the magnitude of the number}
\usage{
choose_scientific(breaks)
}
\arguments{
\item{breaks}{A vector of numerics}
}
\value{
Character vector
}
\description{
Decides whether number should be in normal or scientific form depending on
the magnitude of the number
}
