% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{collect_empirical_island_stacs}
\alias{collect_empirical_island_stacs}
\title{Collect the \code{stac}s of an \code{empirical_island}}
\usage{
collect_empirical_island_stacs(empirical_island)
}
\value{
a numeric vector of \code{stac}s
}
\description{
Collect the \code{stac}s of an \code{empirical_island}
}
\author{
Richèl J.C. Bilderbeek
}
