% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{collect_ideal_island_stacs}
\alias{collect_ideal_island_stacs}
\title{Collect the \code{stac}s of an \code{ideal_island}}
\usage{
collect_ideal_island_stacs(ideal_island)
}
\arguments{
\item{ideal_island}{a simulated island with perfect information,
as created by \link{sim_island} (together with \code{empirical_island},
which has imperfect information).

\code{ideal_island} is a list  containing 3 components:
\itemize{
\item \verb{$branching_times}: island age and stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge and
Endemic anagenetic species.

For cladogenetic species these should
be island age and branching times of the radiation including the
stem age of the radiation.
\item \verb{$stac}: An integer ranging from 1 to 6
indicating the status of the colonist:
\enumerate{
\item Non_endemic_MaxAge
\item Endemic
\item Endemic&Non_Endemic
\item Non_endemic
\item Endemic_singleton_MaxAge
\item Endemic_clade_MaxAge
}
\item \verb{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)
}}
}
\value{
a numeric vector of \code{stac}s
}
\description{
Collect the \code{stac}s of an \code{ideal_island}
}
\author{
Richèl J.C. Bilderbeek
}
