% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_ancestor_time.R
\name{common_ancestor_time}
\alias{common_ancestor_time}
\title{Calculates the branching time of the common ancestor of two species on the
mainland.}
\usage{
common_ancestor_time(total_time, mainland_spec, mainland_clade)
}
\arguments{
\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{mainland_spec}{Numeric focal species on the mainland}

\item{mainland_clade}{Data frame with state of mainland. see
\code{\link[=create_test_mainland_clade]{create_test_mainland_clade()}}.}
}
\value{
Numeric
}
\description{
Calculates the branching time of the common ancestor of two species on the
mainland.
}
\author{
Joshua W. Lambert
}
\keyword{internal}
