% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{count_extant_mainland_species}
\alias{count_extant_mainland_species}
\title{Count the number of extant mainland species}
\usage{
count_extant_mainland_species(mainland)
}
\arguments{
\item{mainland}{the evolutionary history of the mainland species,
as created by \link{sim_mainland}.
Use \link{plot_mainland} to visualise the that evolutionary history.}
}
\value{
the number of extant mainland species
}
\description{
Count the number of extant mainland species
}
\author{
Richèl J.C. Bilderbeek
}
