% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_daisie_data.R
\name{create_daisie_data}
\alias{create_daisie_data}
\title{Converts simulation output into island output}
\usage{
create_daisie_data(
  total_time,
  island_tbl,
  mainland_clade,
  mainland_sample_prob,
  mainland_sample_type
)
}
\arguments{
\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{island_tbl}{Data frame with current state of simulation containing
number of species, see \code{\link[=create_test_island_tbl]{create_test_island_tbl()}}}

\item{mainland_clade}{Data frame with state of mainland, see
\code{\link[=create_test_mainland_clade]{create_test_mainland_clade()}}}

\item{mainland_sample_prob}{Numeric between zero and one determining the
probability of an extant mainland species being sampled.}

\item{mainland_sample_type}{String either "unsampled" for unsampled known
species on the mainland, "undiscovered" for undiscovered species on the
mainland, or "complete" if the mainland sampling probability is one.}
}
\value{
List with the island information, composed of
branching times of extant species, status of species on
the island and number of missing species.
}
\description{
Converts simulation output into island output
}
\author{
Joshua W. Lambert
}
\keyword{internal}
