% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_false_clade_brts.R
\name{create_false_clade_brts}
\alias{create_false_clade_brts}
\title{Creates a vector of branching times when empirical data is a clade but
the ideal data is multiple colonisations}
\usage{
create_false_clade_brts(total_time, anc_branch_t_bp, subset_island)
}
\arguments{
\item{total_time}{Numeric defining the length of the simulation in time units.}

\item{anc_branch_t_bp}{Numeric ancestral branching time before the present
of the immigrant species ancestor and its extant relative on the mainland}

\item{subset_island}{Data frame of island species that originated from a
single colonisation event}
}
\value{
numeric vector
}
\description{
Creates a vector of branching times when empirical data is a clade but
the ideal data is multiple colonisations
}
\keyword{internal}
