% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{create_ihs_breaks}
\alias{create_ihs_breaks}
\title{Creates the axis numbers (breaks) for plotting with an inverse hyperbolic
sine transformation}
\usage{
create_ihs_breaks()
}
\value{
A function that takes a numeric vector
}
\description{
Creates the axis numbers (breaks) for plotting with an inverse hyperbolic
sine transformation
}
