% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_island.R
\name{create_island}
\alias{create_island}
\title{Converts simulation output into island output}
\usage{
create_island(totaltime, island_spec, mainland, mainland_sample_prob)
}
\arguments{
\item{totaltime}{Numeric defining the length of the simulation in time units.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}

\item{mainland}{Data frame with state of mainland. see
\code{\link{create_test_mainland}}.}

\item{mainland_sample_prob}{Numeric between zero and one determining the
probability of an extant mainland species being sampled.}
}
\value{
List with the island information, composed of
branching times of extant species, status of species on
the island and number of missing species.
}
\description{
Converts simulation output into island output
}
\keyword{internal}
