% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{create_plot_breaks}
\alias{create_plot_breaks}
\title{Creates the axis numbers (breaks) for plotting with an inverse hyperbolic
sine transformation, with rounding to a set accuracy to reduce decimal
places plotted}
\usage{
create_plot_breaks(lower_lim, upper_lim, accuracy, round_func)
}
\arguments{
\item{lower_lim}{Numeric stating the lower bound of the axis to be plotted}

\item{upper_lim}{Numeric stating the upper bound of the axis to be plotted}

\item{accuracy}{Numeric stating to which decimal place to kept, e.g. 0.1 is
to one decimal place}

\item{round_func}{Function to be passed through to the rounding, either
\code{floor} or \code{ceiling} function}
}
\value{
Numeric vector
}
\description{
Creates the axis numbers (breaks) for plotting with an inverse hyperbolic
sine transformation, with rounding to a set accuracy to reduce decimal
places plotted
}
