% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{create_plot_labels}
\alias{create_plot_labels}
\title{Creates the axis numbers (labels) for plotting with an inverse hyperbolic
sine transformation, with rounding to a set accuracy to reduce decimal
places plotted}
\usage{
create_plot_labels(lower_lim, upper_lim, accuracy, round_func)
}
\value{
Character vector
}
\description{
Creates the axis numbers (labels) for plotting with an inverse hyperbolic
sine transformation, with rounding to a set accuracy to reduce decimal
places plotted
}
