% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{create_singleton_phylo}
\alias{create_singleton_phylo}
\title{create a singleton}
\usage{
create_singleton_phylo(age)
}
\arguments{
\item{age}{Numeric defining age of singleton species}
}
\description{
create a singleton
}
