% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_island_spec.R
\name{create_test_island_spec}
\alias{create_test_island_spec}
\title{Creates an island community for testing purposes}
\usage{
create_test_island_spec(island_scenario)
}
\arguments{
\item{defaultParamsdoc}{}
}
\value{
data frame
}
\description{
Creates an island community for testing purposes
}
\examples{
island_spec <- DAISIEmainland:::create_test_island_spec(island_scenario = 1)
}
\keyword{internal}
