% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_island_tbl.R
\name{create_test_island_tbl}
\alias{create_test_island_tbl}
\title{Creates an island community for testing purposes}
\usage{
create_test_island_tbl(island_scenario)
}
\value{
data frame with island community information
}
\description{
Creates an island community for testing purposes
}
\author{
Joshua W. Lambert
}
\keyword{internal}
