% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_mainland.R
\name{create_test_mainland}
\alias{create_test_mainland}
\title{Creates a clade on the mainland for testing purposes}
\usage{
create_test_mainland(mainland_scenario)
}
\arguments{
\item{mainland_scenario}{Integer determining which mainland clade scenario
is produced}
}
\value{
data frame
}
\description{
Creates a clade on the mainland for testing purposes
}
