% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_data_to_tables.R
\name{daisie_data_colonist_info_to_braching_times_table}
\alias{daisie_data_colonist_info_to_braching_times_table}
\title{Internal function}
\usage{
daisie_data_colonist_info_to_braching_times_table(daisie_data_colonist_info)
}
\arguments{
\item{daisie_data_colonist_info}{an element of a \code{daisie_data},
that is not the first element (the first element is of type
\code{daisie_data_header}).}
}
\value{
a table with as much rows as branching times
}
\description{
Convert the non-header part of a \code{daisie_data} into a tabular format,
contain per colonist the branching times. This excludes the time
of colonisation.
}
\seealso{
use \link{daisie_data_colonist_info_to_colonisation_times_table}
to obtain the colonisation times of the colonists
}
\author{
Richèl J.C. Bilderbeek
}
