% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_data_to_tables.R
\name{daisie_data_colonist_info_to_general_table}
\alias{daisie_data_colonist_info_to_general_table}
\title{Convert the header of a \code{daisie_data} into a tabular format}
\usage{
daisie_data_colonist_info_to_general_table(daisie_data_colonist_info)
}
\arguments{
\item{daisie_data_colonist_info}{an element of a \code{daisie_data},
that is not the first element (the first element is of type
\code{daisie_data_header}).}
}
\value{
a table with one row,
containing all elements in a \code{daisie_data_colonist_info} header
that are not branching times.
}
\description{
Convert the header of a \code{daisie_data} into a tabular format
}
\author{
Richèl J.C. Bilderbeek
}
