% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_data_to_tables.R
\name{daisie_data_colonist_info_to_table}
\alias{daisie_data_colonist_info_to_table}
\title{Convert the header of a \code{daisie_data} into a tabular format}
\usage{
daisie_data_colonist_info_to_table(daisie_data_colonist_info)
}
\value{
a table with one row,
containing all elements in a \code{daisie_data} header
}
\description{
Convert the header of a \code{daisie_data} into a tabular format
}
\note{
the header of a \code{daisie_data} is its first element
}
\author{
Richèl J.C. Bilderbeek
}
