% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_data_to_tables.R
\name{daisie_data_to_tables}
\alias{daisie_data_to_tables}
\title{Convert an \code{daisie_data} to a list of tables}
\usage{
daisie_data_to_tables(daisie_data)
}
\arguments{
\item{daisie_data}{\link{list} containing data of of a DAISIE simulation
with mainland dynamics,
as produced by \link{sim_island_with_mainland}
and checked by \link{check_daisie_data}

A \code{daisie_data} has two elements:
\itemize{
\item \code{ideal_islands}: the ideal data set produced in a simulation,
see \link{check_ideal_daisie_data} for more details
\item \code{empirical_islands}: the empirical data sets produced in a simulation
see \link{check_empirical_daisie_data} for more details
}}
}
\value{
a \link{list} with elements:
\itemize{
\item \code{ideal_island_table}: the \code{daisie_data$empirical_islands} as a table,
as done by \link{empirical_daisie_data_to_tables}
\item \code{empirical_island_table}: the \code{daisie_data$ideal_island} as a table
as done by \link{ideal_daisie_data_to_tables}
}
}
\description{
Convert an \code{daisie_data} to a list of tables
}
\examples{
set.seed(
  4,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
daisie_data <- sim_island_with_mainland(
  total_time = 1.0,
  m = 10,
  island_pars = c(1, 1, 10, 0.1, 1),
  mainland_ex = 1,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete",
  replicates = 1)
daisie_data_to_tables(daisie_data)

}
\author{
Richèl J.C. Bilderbeek
}
