% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_data_to_tables.R
\name{daisie_header_to_table}
\alias{daisie_header_to_table}
\title{Convert a \code{daisie_data_header} into a tabular format}
\usage{
daisie_header_to_table(daisie_data_header)
}
\arguments{
\item{daisie_data_header}{the first element of a \code{daisie_data}}
}
\value{
a table with one row,
containing all elements in a \code{daisie_data_header}
}
\description{
Convert a \code{daisie_data_header} into a tabular format
}
\author{
Richèl J.C. Bilderbeek
}
