% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_to_tables.R
\name{empirical_island_to_tables}
\alias{empirical_island_to_tables}
\title{Convert an \code{empirical_island} to a list of tables}
\usage{
empirical_island_to_tables(empirical_island)
}
\arguments{
\item{empirical_island}{a simulated island with imperfect information,
with one \code{empirical_island_clade} per simulated island clade.

An \code{empirical_island} is created by \link{sim_island}
(together with \code{ideal_island}, which has perfect information),
as can be checked by \link{check_empirical_island}.}
}
\value{
a \link{list} with elements:
\itemize{
\item \code{speciations}: a table with the speciation events
\item \code{colonisations}: a table with colonisations
}
}
\description{
Convert an \code{empirical_island} to a list of tables
}
