% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_to_str.R
\name{event_to_str}
\alias{event_to_str}
\title{Convert an event to a string}
\usage{
event_to_str(possible_event)
}
\arguments{
\item{possible_event}{Numeric defining what event will happen.}
}
\value{
the event as a string
}
\description{
Convert an event to a string
}
\examples{
event_to_str(1)
event_to_str(2)
event_to_str(3)
event_to_str(4)

}
\seealso{
use \link{str_to_event} to convent a string to an \code{event}
}
\author{
Richel J.C. Bilderbeek
}
\keyword{internal}
