% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Formats simulation output into standard DAISIE list output.}
\usage{
format_data(island_replicates, time, m, verbose = TRUE)
}
\arguments{
\item{island_replicates}{List that has as many elements as replicates. Each
element must be a list with the elements \code{island_age} and
\code{not_present}. ##### LOOK INTO THIS}

\item{time}{Numeric defining the length of the simulation in time units.}

\item{m}{Numeric defining the size of mainland pool.}

\item{verbose}{Logical, determining if progress output should be printed
during the simulation.}
}
\value{
List with DAISIE simulation output
}
\description{
Formats simulation output into standard DAISIE list output.
}
\keyword{internal}
