% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_to_daisie_data.R
\name{format_to_daisie_data}
\alias{format_to_daisie_data}
\title{Formats simulation output into standard DAISIE list output.}
\usage{
format_to_daisie_data(island_replicates, total_time, m)
}
\arguments{
\item{island_replicates}{a \link{list} of that has as many elements as
replicates. Each element is a \link{list} of \code{islands} (as checked
by \link{check_island}).
Use \link{check_island_replicates} to check a \code{island_replicates}
being correct}

\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{m}{Numeric defining the size of mainland pool.}
}
\value{
a \code{daisie_data} object,
as can be checked by \link{check_daisie_data}
}
\description{
Formats simulation output into standard DAISIE list output.
}
\author{
Joshua W. Lambert
}
\keyword{internal}
