% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_data_to_tables.R
\name{ideal_daisie_data_to_tables}
\alias{ideal_daisie_data_to_tables}
\title{Convert an \code{ideal_island} to a list of tables}
\usage{
ideal_daisie_data_to_tables(ideal_daisie_data)
}
\value{
a \link{list} with elements:
\itemize{
\item \code{speciations}: a table with the speciation events
\item \code{colonisations}: a table with colonisations
}
}
\description{
Convert an \code{ideal_island} to a list of tables
}
