% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_daisie_datalist.R
\name{is_daisie_data}
\alias{is_daisie_data}
\title{Determine if a \code{daisie_data} is valid.}
\usage{
is_daisie_data(daisie_data, verbose = FALSE)
}
\arguments{
\item{daisie_data}{\link{list} containing data of of a DAISIE simulation
with mainland dynamics,
as produced by \link{sim_island_with_mainland}
and checked by \link{check_daisie_data}

A \code{daisie_data} has two elements:
\itemize{
\item \code{ideal_islands}: the ideal data set produced in a simulation,
see \link{check_ideal_daisie_data} for more details
\item \code{empirical_islands}: the empirical data sets produced in a simulation
see \link{check_empirical_daisie_data} for more details
}}

\item{verbose}{Logical, determining if progress output should be printed
during the simulation.}
}
\value{
\link{TRUE} if the \code{daisie_data} is valid.
}
\description{
Determine if a \code{daisie_data} is valid.
}
\author{
Richèl J.C. Bilderbeek
}
