% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_to_phylos.R
\name{island_to_phylos}
\alias{island_to_phylos}
\title{Converts a DAISIE datalist (either ideal or empirical) to a list of phylo
objects that can be used to plot the island.}
\usage{
island_to_phylos(island)
}
\arguments{
\item{island}{The history of a a single island,
as can be checked by \link{check_island}.

An \code{island} is \link{list} with two elements:
\itemize{
\item \code{ideal_island}: the island history based on ideal data,
as can be checked by \link{check_ideal_island}
\item \code{empirical_island}: the island history based on empirical data
as can be checked by \link{check_empirical_island}
For recolonising lineages, there is an extra element,
\code{all_colonisations} per list element.
It is comprised of \verb{$event_times} and \verb{$species_type}:
\describe{
\item{\verb{$event_times}}{ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times. Each vector pertains to one
colonising lineage.}
\item{\verb{$species_type}}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}
}}
}
\value{
List of phylo objects. See ape R package for details on phylo class
}
\description{
Converts a DAISIE datalist (either ideal or empirical) to a list of phylo
objects that can be used to plot the island.
}
