% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_to_tables.R
\name{island_to_tables}
\alias{island_to_tables}
\title{Convert an \code{island} to a list of tables}
\usage{
island_to_tables(island)
}
\arguments{
\item{island}{The history of a a single island,
as can be checked by \link{check_island}.

An \code{island} is \link{list} with two elements:
\itemize{
\item \code{ideal_island}: the island history based on ideal data,
as can be checked by \link{check_ideal_island}
\item \code{empirical_island}: the island history based on empirical data
as can be checked by \link{check_empirical_island}
}}
}
\value{
a \link{list} with elements:
\itemize{
\item \code{speciations}: a table with the speciation events
\item \code{colonisations}: a table with colonisations
}
}
\description{
Convert an \code{island} to a list of tables
}
\author{
Richèl J.C. Bilderbeek
}
