% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctt.R
\name{plot_ctt}
\alias{plot_ctt}
\title{Plots a heatmap of the colonisation through time statistic for mainland
sampling probability (y-axis) and mainland extinction rate (x-axis).}
\usage{
plot_ctt(save = TRUE, test = FALSE)
}
\arguments{
\item{save}{Boolean determining whether a plot is saved (TRUE) or returned
(FALSE)}

\item{test}{Boolean determining whether a function is being tested}
}
\value{
Void (saves plot)
}
\description{
Plots a heatmap of the colonisation through time statistic for mainland
sampling probability (y-axis) and mainland extinction rate (x-axis).
}
