% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_daisie_data.R
\name{plot_daisie_data}
\alias{plot_daisie_data}
\title{Plot the simulated island, based on ideal and empirical data.}
\usage{
plot_daisie_data(daisie_data)
}
\arguments{
\item{daisie_data}{List containing data of DAISIE simulation with mainland
dynamics. Output from `sim_island_with_mainland``}
}
\value{
a \code{ggplot2}
}
\description{
Plot the simulated island, based on ideal and empirical data.
}
\details{
The plot will consist of a grid,
with two columns and one or more rows.
The columns depict the same evolutionary history of the
island, yet for different type of data.
The rows depict the species clades.

The vertical lines in the column for the ideal data depict
one or more colonisation events.
}
\examples{
set.seed(
  9,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
total_time <- 1
m <- 10 # number of mainland clades
mainland <- sim_mainland(
  total_time = total_time,
  m = m,
  mainland_ex = 2.0
)
mainland_clade <- mainland[[1]]
plot_mainland_clade(mainland_clade)
island <- DAISIEmainland:::sim_island(
  total_time = total_time,
  island_pars = c(1, 1, 10, 12, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
plot_island(island)
daisie_data <- DAISIEmainland:::format_to_daisie_data(
  island_replicates = island,
  total_time = total_time,
  m = m
)
plot_daisie_data(daisie_data)
}
\seealso{
These are the functions to plot an evolutionary history:
\itemize{
\item Use \link{plot_mainland} to plot the mainland history
\item Use \link{plot_island} to plot the island history.
based on both empirical data and ideal data.
\item Use \link{plot_empirical_island} to plot the island history.
based on empirical data.
\item Use \link{plot_ideal_island} to plot the island history
based on ideal data.
}
}
\author{
Richèl J.C. Bilderbeek
}
