% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_empirical_island.R
\name{plot_empirical_island}
\alias{plot_empirical_island}
\title{Plot the simulated ideal island.}
\usage{
plot_empirical_island(empirical_island)
}
\arguments{
\item{empirical_island}{a simulated island with imperfect information,
with one \code{empirical_island_clade} per simulated island clade.

An \code{empirical_island} is created by \link{sim_island}
(together with \code{ideal_island}, which has perfect information),
as can be checked by \link{check_empirical_island}.}
}
\value{
a \code{ggplot2}
}
\description{
Plot the simulated ideal island.
}
\examples{
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = 2)
island <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 1, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
plot_empirical_island(island$empirical_island)
}
\seealso{
These are the functions to plot an evolutionary history:
\itemize{
\item Use \link{plot_mainland} to plot the mainland history
\item Use \link{plot_island} to plot the island history.
based on both empirical data and ideal data.
\item Use \link{plot_empirical_island} to plot the island history.
based on empirical data.
\item Use \link{plot_ideal_island} to plot the island history
based on ideal data.
}
}
\author{
Richèl J.C. Bilderbeek
}
