% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ideal_island.R
\name{plot_ideal_island}
\alias{plot_ideal_island}
\title{Plot the simulated ideal island.}
\usage{
plot_ideal_island(ideal_island)
}
\arguments{
\item{ideal_island}{a simulated island with perfect information,
which is a \link{list} of one or more \link{ideal_island_clade}s.

An \code{ideal_island} is created by \link{sim_island} (together with \code{empirical_island},
which has imperfect information),
as can be checked by \link{check_ideal_island}}
}
\value{
a \code{ggplot2}
}
\description{
Plot the simulated ideal island.
}
\examples{
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = 2)
island <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 1, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
plot_ideal_island(island$ideal_island)

}
\seealso{
These are the functions to plot an evolutionary history:
\itemize{
\item Use \link{plot_mainland} to plot the mainland history
\item Use \link{plot_island} to plot the island history.
based on both empirical data and ideal data.
\item Use \link{plot_empirical_island} to plot the island history.
based on empirical data.
\item Use \link{plot_ideal_island} to plot the island history
based on ideal data.
}
}
\author{
Richèl J.C. Bilderbeek
}
