% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ihs.R
\name{plot_ihs}
\alias{plot_ihs}
\title{Plots the true value versus the inverse hyperbolic sine tranformed value to
show the relationship.}
\usage{
plot_ihs(output_file_path)
}
\arguments{
\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}
}
\value{
Void (saves plot)
}
\description{
Plots the true value versus the inverse hyperbolic sine tranformed value to
show the relationship.
}
