% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inf_k.R
\name{plot_inf_k}
\alias{plot_inf_k}
\title{Plots the percentage of carrying capacity (K') estimates that are infinite}
\usage{
plot_inf_k(data_folder_path, output_file_path)
}
\arguments{
\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}
}
\value{
Void (saves plot)
}
\description{
Plots the percentage of carrying capacity (K') estimates that are infinite
}
