% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_island.R
\name{plot_island}
\alias{plot_island}
\title{Plot the simulated island, based on ideal and empirical data.}
\usage{
plot_island(island)
}
\arguments{
\item{island}{A single island replicate, can be either ideal or empirical
data}
}
\value{
a \code{ggplot2}
}
\description{
Plot the simulated island, based on ideal and empirical data.
}
\examples{
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = mainland_scenario
)
island <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 12, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
plot_island(island)

}
\seealso{
These are the functions to plot an evolutionary history:
\itemize{
\item Use \link{plot_mainland} to plot the mainland history
\item Use \link{plot_island} to plot the island history.
based on both empirical data and ideal data.
\item Use \link{plot_empirical_island} to plot the island history.
based on empirical data.
\item Use \link{plot_ideal_island} to plot the island history
based on ideal data.
}
}
\author{
Richèl J.C. Bilderbeek
}
