% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_island_phylos.R
\name{plot_island_phylos}
\alias{plot_island_phylos}
\title{Plots island phylogenetic data}
\usage{
plot_island_phylos(island)
}
\arguments{
\item{island_phylos}{stub}

\item{island_age}{stub}
}
\value{

}
\description{
Plots island phylogenetic data
}
