% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_island_tbl.R
\name{plot_island_tbl}
\alias{plot_island_tbl}
\title{Plot the simulated island, based on ideal and empirical data.}
\usage{
plot_island_tbl(total_time, island_tbl)
}
\arguments{
\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{island_tbl}{Data frame with current state of simulation containing
number of species, see \code{\link[=create_test_island_tbl]{create_test_island_tbl()}}}
}
\value{
a \code{ggplot2}
}
\description{
Plot the simulated island, based on ideal and empirical data.
}
\details{
The plot will consist of a grid,
with two columns and one or more rows.
The columns depict the same evolutionary history of the
island, yet for different type of data.
The rows depict the species clades.

The vertical lines in the column for the ideal data depict
one or more colonisation events.
}
\examples{
\dontrun{
set.seed(
  9,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
mainland <- sim_mainland(
  total_time = 10,
  m = 10,
  mainland_ex = 1.0
)
mainland_clade <- mainland[[1]]
plot_mainland_clade(mainland_clade)
island_tbl <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 12, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
plot_island_tbl(island_tbl)
}
}
\seealso{
These are the functions to plot an evolutionary history:
\itemize{
\item Use \link{plot_mainland} to plot the mainland history
\item Use \link{plot_island_tbl} to plot the island history.
based on both empirical data and ideal data.
}
}
\author{
Richèl J.C. Bilderbeek
}
