% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_k_estimates.R
\name{plot_k_estimates}
\alias{plot_k_estimates}
\title{Plots carrying capacity estimates}
\usage{
plot_k_estimates(data_folder_path, output_file_path, param_space)
}
\arguments{
\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{param_space}{String specifying which parameter space results to read.
Either "general", "mainland_ex", or "mainland_sample_prob"}
}
\value{
Void (saves plot)
}
\description{
Plots carrying capacity estimates
}
