% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log_param_estimates.R
\name{plot_log_param_estimates}
\alias{plot_log_param_estimates}
\title{Plots a grid of density plots of parameter estimates on the diagonal and
scatter plots of parameter estimates on a log scale under the diagonal.}
\usage{
plot_log_param_estimates(
  param_set,
  data_folder_path,
  output_file_path,
  parameter,
  num_breaks,
  signif
)
}
\arguments{
\item{param_set}{Numeric for the index of the parameter set from the
parameter space}

\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{parameter}{String specifying which parameter is plotted.
"mainland_ex" plots mainland extinction as the variable,
"mainland_sample_prob" plots mainland sampling probability as the variable
and "both" plots all the data}

\item{num_breaks}{A numeric specifying how many breaks are wanted on the
plot axes}

\item{signif}{A numeric specifying how many significant figures the axes
labels have when plotting}
}
\value{
Void (saves plot)
}
\description{
Plots a grid of density plots of parameter estimates on the diagonal and
scatter plots of parameter estimates on a log scale under the diagonal.
}
