% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mainland.R
\name{plot_mainland}
\alias{plot_mainland}
\title{Plot the simulated mainland}
\usage{
plot_mainland(mainland)
}
\arguments{
\item{mainland}{the evolutionary history of the mainland species,
as created by \link{sim_mainland}}
}
\value{
a \code{ggplot2}
}
\description{
Plot the simulated mainland
}
\examples{
mainland <- DAISIEmainland:::sim_mainland(
  total_time = 1,
  m = 2,
  mainland_ex = 2)

plot_mainland(mainland)

}
\author{
Richèl J.C. Bilderbeek
}
