% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mainland_clade.R
\name{plot_mainland_clade}
\alias{plot_mainland_clade}
\title{Plot the simulated mainland.}
\usage{
plot_mainland_clade(mainland_clade)
}
\arguments{
\item{mainland_clade}{Data frame with state of mainland, see
\code{\link[=create_test_mainland_clade]{create_test_mainland_clade()}}}
}
\value{
a \code{ggplot2}
}
\description{
Plot the simulated mainland.
}
\examples{
mainland_clade <- create_test_mainland_clade(mainland_scenario = 2)
plot_mainland_clade(mainland_clade)
}
\author{
Richèl J.C. Bilderbeek
}
