% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_max_age_sup.R
\name{plot_max_age_sup}
\alias{plot_max_age_sup}
\title{Plots two faceted violin plots of the percent of max age colonists on the
island for different values of mainland extinction. The left side facet has
the ideal data and the right side facet has the empirical data.}
\usage{
plot_max_age_sup(data_folder_path, output_file_path, param_space)
}
\arguments{
\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{param_space}{String specifying which parameter space results to read.
Either "general", "mainland_ex", or "mainland_sample_prob"}
}
\value{
Plot
}
\description{
Plots two faceted violin plots of the percent of max age colonists on the
island for different values of mainland extinction. The left side facet has
the ideal data and the right side facet has the empirical data.
}
