% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_param_diffs.R
\name{plot_param_diffs}
\alias{plot_param_diffs}
\title{Plots a grid of density plots of parameter estimates on the diagonal,
scatter plots of parameter estimates under the diagonal and scatter plots
of differences between ideal and empirical parameter estimates above the
diagonal.}
\usage{
plot_param_diffs(
  param_set,
  data_folder_path,
  output_file_path,
  parameter,
  signif,
  scientific = FALSE,
  transform = NULL
)
}
\arguments{
\item{param_set}{Numeric for the index of the parameter set from the
parameter space}

\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{parameter}{String specifying which parameter is plotted.
"mainland_ex" plots mainland extinction as the variable,
"mainland_sample_prob" plots mainland sampling probability as the variable
and "both" plots all the data}

\item{signif}{A numeric specifying how many significant figures the axes
labels have when plotting}

\item{scientific}{A boolean determining whether the axis labels will be
converted to scientific notation}

\item{transform}{Either \code{NULL} or \code{"ihs"} to specify no transformation of
the data or an inverse hyperbolic sine (ihs) transformation}
}
\value{
Void (saves plot)
}
\description{
Plots a grid of density plots of parameter estimates on the diagonal,
scatter plots of parameter estimates under the diagonal and scatter plots
of differences between ideal and empirical parameter estimates above the
diagonal.
}
