% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_metrics.R
\name{plot_sim_metrics}
\alias{plot_sim_metrics}
\title{Plots metrics from simulations}
\usage{
plot_sim_metrics(analysis_results, output_file_path)
}
\arguments{
\item{analysis_results}{A list of analysis result elements which can be
checked with \code{check_analysis_result}}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}
}
\value{
none Four plots are shown: histogram of the number of species,
histogram of the number of colonizations, histogram of the largest clade size
and histogram of the rank of the largest clade
}
\description{
Plots metrics from simulations
}
\author{
Joshua W. Lambert
}
