% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_analysis_results.R
\name{read_analysis_results}
\alias{read_analysis_results}
\title{Reads all the the results from a directory, with path is specified in
\code{data_folder_path}.}
\usage{
read_analysis_results(data_folder_path)
}
\arguments{
\item{data_folder_path}{String specifying the directory the data is read
from}
}
\value{
A list of results
}
\description{
Reads all the the results from a directory, with path is specified in
\code{data_folder_path}.
}
\examples{
results <- DAISIEmainland::read_analysis_results(
  file.path("tests", "testthat", "testdata")
)
}
