% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_event.R
\name{sample_event}
\alias{sample_event}
\title{Samples what event to happen next on the island.}
\usage{
sample_event(rates)
}
\arguments{
\item{rates}{named list of numeric rates as returned by
\code{\link{update_rates}()}.}
}
\value{
numeric indicating what event will happen, or a supposed event that
would happen in some timesteps of the ontogeny algorithm.
\enumerate{
\item immigration event
\item extinction event
\item cladogenesis event
\item anagenesis event
}
}
\description{
Samples what event to happen next on the island.
}
\author{
Joshua W. Lambert
}
\keyword{internal}
