% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{scientific}
\alias{scientific}
\title{Creates the axis numbers (labels) for plotting with scientific form in
x10 notation}
\usage{
scientific(breaks)
}
\arguments{
\item{breaks}{A vector of numerics}
}
\value{
Character vector
}
\description{
Creates the axis numbers (labels) for plotting with scientific form in
x10 notation
}
