% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_island_and_mainland.R
\name{sim_island_and_mainland}
\alias{sim_island_and_mainland}
\title{Simulate both the island and mainland,
a version of \link{sim_island_with_main} used in debugging, that
\itemize{
\item returns both the island and the mainland,
to verify the correctness of the simulation
\item has no 1,
to simplify reading
}}
\usage{
sim_island_and_mainland(
  total_time,
  m,
  island_pars,
  mainland_ex,
  mainland_sample_prob,
  mainland_sample_type,
  verbose = FALSE
)
}
\arguments{
\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{m}{Numeric defining the size of mainland pool.}

\item{island_pars}{A numeric vector containing the parameters for the island:
\itemize{
\item{\code{island_pars[1]}: lambda^c (cladogenesis rate)}
\item{\code{island_pars[2]}: mu (extinction rate)}
\item{\code{island_pars[3]}: K (carrying capacity), set K=Inf for
diversity independence.}
\item{\code{island_pars[4]}: gamma (immigration rate)}
\item{\code{island_pars[5]}: lambda^a (anagenesis rate)}
}}

\item{mainland_ex}{Numeric parameter for mainland extinction rate.}

\item{mainland_sample_prob}{Numeric between zero and one determining the
probability of an extant mainland species being sampled.}

\item{mainland_sample_type}{String either "unsampled" for unsampled known
species on the mainland, "undiscovered" for undiscovered species on the
mainland, or "complete" if the mainland sampling probability is one.}

\item{verbose}{Logical, determining if progress output should be printed
during the simulation.}
}
\value{
a list with elements
\itemize{
\item \code{mainland}
\item \code{ideal_island}
\item \code{empirical_island}
}

See \link{sim_island_with_mainland} for the superior documentation
}
\description{
Simulate both the island and mainland,
a version of \link{sim_island_with_main} used in debugging, that
\itemize{
\item returns both the island and the mainland,
to verify the correctness of the simulation
\item has no 1,
to simplify reading
}
}
\examples{
island <- sim_island_and_mainland(
  total_time = 1,
  m = 2, # Number of mainland clades
  island_pars = c(1, 1, 10, 0.1, 1),
  mainland_ex = 1,
  mainland_sample_prob = 1,
  mainland_sample_type = "undiscovered"
)
}
\author{
original function by Joshua W. Lambert,
after which Richèl J.C. Bilderbeek adapted it to this debug function by
}
