% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{singleton_to_segment}
\alias{singleton_to_segment}
\title{Converts singleton phylo object into segment that can be plotted}
\usage{
singleton_to_segment(singleton_phylos)
}
\arguments{
\item{singleton_phylo}{}
}
\value{

}
\description{
Converts singleton phylo object into segment that can be plotted
}
