% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{species_type_to_str}
\alias{species_type_to_str}
\title{Convert an \code{species_type} to a string,
where the \code{species_type} denotes the origin of a species.}
\usage{
species_type_to_str(species_type)
}
\value{
the species type as a string
}
\description{
Convert an \code{species_type} to a string,
where the \code{species_type} denotes the origin of a species.
}
\examples{
species_type_to_str("A") # Anagenetic
species_type_to_str("C") # Cladogentic
species_type_to_str("I") # Immigrant

}
\author{
Richèl J.C. Bilderbeek
}
