% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_to_english_str.R
\name{stac_to_english_str}
\alias{stac_to_english_str}
\title{Convert a \code{stac} (status of colonist) to an English/human-readable string}
\usage{
stac_to_english_str(stac)
}
\arguments{
\item{stac}{the status of the colonist,
see \link{stac_to_str} for values and meanings}
}
\value{
the status of a colonist as a string
}
\description{
Convert a \code{stac} (status of colonist) to an English/human-readable string:
}
\details{
\tabular{ll}{
   Value \tab String \cr
   0 \tab \verb{[no colonization has taken place]} \cr
   1 \tab \verb{Non-endemic at maximum age} \cr
   2 \tab \code{Endemic} \cr
   3 \tab \verb{Endemic and non-endemic} \cr
   4 \tab \code{Non-endemic} \cr
   5 \tab \verb{Endemic singleton at maximum age} \cr
   6 \tab \verb{Endemic clade at maximum age} \cr
}
}
\examples{
DAISIEmainland:::stac_to_english_str(0)
DAISIEmainland:::stac_to_english_str(1)
DAISIEmainland:::stac_to_english_str(2)
DAISIEmainland:::stac_to_english_str(3)
DAISIEmainland:::stac_to_english_str(4)
DAISIEmainland:::stac_to_english_str(5)
DAISIEmainland:::stac_to_english_str(6)
}
\seealso{
use \link{stac_to_str} to convent a \code{stac} to a single-word
computery string
}
\author{
Richèl J.C. Bilderbeek
}
\keyword{internal}
