% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{stac_to_str}
\alias{stac_to_str}
\title{Convert a \code{stac} (status of colonist) to a string}
\usage{
stac_to_str(stac)
}
\arguments{
\item{stac}{the status of the colonist,
see \link{stac_to_str} for values and meanings}
}
\value{
the status of a colonist as a string
}
\description{
Convert a \code{stac} (status of colonist) to a string:
}
\details{
\tabular{ll}{
   Value \tab String \cr
   0 \tab \verb{[no colonization has taken place]} \cr
   1 \tab \code{Non_endemic_MaxAge} \cr
   2 \tab \code{Endemic} \cr
   3 \tab \code{Endemic&Non_Endemic} \cr
   4 \tab \code{Non_endemic} \cr
   5 \tab \code{Endemic_singleton_MaxAge} \cr
   6 \tab \code{Endemic_clade_MaxAge} \cr
}
}
\examples{
DAISIEmainland:::stac_to_str(0)
DAISIEmainland:::stac_to_str(1)
DAISIEmainland:::stac_to_str(2)
DAISIEmainland:::stac_to_str(3)
DAISIEmainland:::stac_to_str(4)
DAISIEmainland:::stac_to_str(5)
DAISIEmainland:::stac_to_str(6)
}
\seealso{
use \link{str_to_stac} to convent a string to a \code{stac}
}
\author{
Richèl J.C. Bilderbeek
}
\keyword{internal}
