% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_event.R
\name{str_to_event}
\alias{str_to_event}
\title{Convert a string to an event}
\usage{
str_to_event(event_str)
}
\arguments{
\item{event_str}{the event as a string}
}
\value{
the event as a string
}
\description{
Convert a string to an event
}
\examples{
str_to_event("immigration")
str_to_event("extinction")
str_to_event("anagenesis")
str_to_event("cladogenesis")

}
\seealso{
use \link{event_to_str} to convent an \code{event} to a string
}
\author{
Richel J.C. Bilderbeek
}
\keyword{internal}
