% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_stac.R
\name{str_to_stac}
\alias{str_to_stac}
\title{Convert a string to a \code{stac}}
\usage{
str_to_stac(stac_str)
}
\arguments{
\item{stac_str}{the \code{stac} as a string}
}
\value{
the \code{stac} as a string
}
\description{
Convert a string to a \code{stac}
}
\examples{
DAISIEmainland:::str_to_stac("Non_endemic_MaxAge")
DAISIEmainland:::str_to_stac("Endemic")
DAISIEmainland:::str_to_stac("Endemic&Non_Endemic")
DAISIEmainland:::str_to_stac("Non_endemic_MaxAge")
DAISIEmainland:::str_to_stac("Endemic_singleton_MaxAge")
DAISIEmainland:::str_to_stac("Endemic_clade_MaxAge")

}
\seealso{
use \link{stac_to_str} to convent an \code{stac} to a string
}
\author{
Richèl J.C. Bilderbeek
}
\keyword{internal}
