% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_state.R
\name{update_state}
\alias{update_state}
\title{Updates state of island given sampled event for a constant rate case.}
\usage{
update_state(
  timeval,
  total_time,
  possible_event,
  max_spec_id,
  mainland_spec,
  island_spec
)
}
\arguments{
\item{timeval}{Numeric defining current time of simulation.}

\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{possible_event}{Numeric defining what event will happen.}

\item{max_spec_id}{Current species IDs.}

\item{mainland_spec}{Numeric focal species on the mainland}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}
}
\value{
The updated state of the system, which is a list with the
\code{island_spec} matrix and an integer \code{max_spec_id} with the most
recent ID of species.
}
\description{
Makes the event happen by updating island species matrix and species IDs.
What event happens is determined by the sampling in the algorithm.
}
\author{
Joshua W. Lambert
}
\keyword{internal}
