% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_colonisations_to_table.R
\name{all_colonisations_to_table}
\alias{all_colonisations_to_table}
\title{Extract all colonisations from an island history
and put these in a table}
\usage{
all_colonisations_to_table(ideal_or_empirical_island)
}
\arguments{
\item{ideal_or_empirical_island}{the evolutionary history of
an island from ideal or empirical data,
as created by \link{sim_island}

\code{ideal_or_empirical_island} is a list  containing 3 components:
\itemize{
\item \verb{$branching_times}: island age and stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge and
Endemic anagenetic species.

For cladogenetic species these should
be island age and branching times of the radiation including the
stem age of the radiation.
\item \verb{$stac}: An integer ranging from 1 to 6
indicating the status of the colonist:
\enumerate{
\item Non_endemic_MaxAge
\item Endemic
\item Endemic&Non_Endemic
\item Non_endemic
\item Endemic_singleton_MaxAge
\item Endemic_clade_MaxAge
}
\item \verb{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)
}

For recolonising lineages, there is an extra element,
\code{all_colonisations} per list element.
It is comprised of \verb{$event_times} and \verb{$species_type}:
\describe{
\item{\verb{$event_times}}{ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times. Each vector pertains to one
colonising lineage.}
\item{\verb{$species_type}}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}}
}
\value{
a \link{list} with elements:
\itemize{
\item \code{t}: a table
\item \code{empirical_island$all_colonisations}: a list with all colonizations
in the empirical data
\item \code{ideal_island$all_colonisations}: a list with all colonizations
in the ideal data
}
}
\description{
Extract all colonisations from an island history
and put these in a table
}
\author{
Richèl J.C. Bilderbeek
}
