% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_endemic_singletons.R
\name{all_endemic_singletons}
\alias{all_endemic_singletons}
\title{Checks if all island species are endemic singletons. Singletons are single
lineages that have not formed an island clade of two or more species.}
\usage{
all_endemic_singletons(daisie_data)
}
\arguments{
\item{daisie_data}{\link{list} containing data of of a DAISIE simulation
with mainland dynamics,
as produced by \link{sim_island_with_mainland}
and checked by \link{check_daisie_data}

A \code{daisie_data} has two elements:
\itemize{
\item \code{ideal_islands}: the ideal data set produced in a simulation,
see \link{check_ideal_daisie_data} for more details
\item \code{empirical_islands}: the empirical data sets produced in a simulation
see \link{check_empirical_daisie_data} for more details
}}
}
\value{
Boolean
}
\description{
Checks if all island species are endemic singletons. Singletons are single
lineages that have not formed an island clade of two or more species.
}
\examples{
\dontrun{
island <- DAISIEmainland::sim_island_with_mainland(
  total_time = 1,
  m = 100,
  island_pars = c(1,1,10,0.1,1),
  mainland_ex = 0.1,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete",
  replicates = 2,
  verbose = FALSE)
bool <- all_endemic_singletons(island$ideal_islands[[1]])
}
}
\author{
Joshua W. Lambert
}
