% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_endemic_singletons.R
\name{all_endemic_singletons}
\alias{all_endemic_singletons}
\title{Checks if all island species are endemic singletons. Singletons are single
lineages that have not formed an island clade of two or more species.}
\usage{
all_endemic_singletons(island)
}
\arguments{
\item{island}{A single island replicate, can be either ideal or empirical
data}
}
\value{
Boolean
}
\description{
Checks if all island species are endemic singletons. Singletons are single
lineages that have not formed an island clade of two or more species.
}
\examples{
\dontrun{
island <- DAISIEmainland::sim_island_with_mainland(
  total_time = 1,
  m = 100,
  island_pars = c(1,1,10,0.1,1),
  mainland_ex = 0.1,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete",
  replicates = 2,
  verbose = FALSE)
bool <- all_endemic_singletons(island$ideal_islands[[1]])
}
}
\author{
Joshua W. Lambert
}
