% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_code_to_y.R
\name{branch_code_to_y}
\alias{branch_code_to_y}
\title{Convert a \code{branch_code} to a y coordinat in range \link{0, 1}}
\usage{
branch_code_to_y(branch_code)
}
\arguments{
\item{branch_code}{a species' branching pattern,
as checked by \link{check_branch_code}, e.g. \code{A}, \code{AB}, \code{AABA}.}
}
\value{
a y coordinat in range \link{0, 1}
}
\description{
Convert a \code{branch_code} to a y coordinat in range \link{0, 1}.
}
\details{
The y coordinat is the result of a division.
The numerator in this division equals the binary value of the
\code{branch_code} (where the \code{A} equals a zero, and the \code{B} equals a
one) plus one.
The denominator in this division equals the number of characters in
the \code{branch_code} plus one.
}
\examples{
branch_code_to_y("A")
branch_code_to_y("AA")
branch_code_to_y("AB")
}
\author{
Richèl J.C. Bilderbeek
}
