% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_code_to_y.R
\name{branch_code_to_y}
\alias{branch_code_to_y}
\title{Convert a \code{branch_code} to a y coordinat in range \verb{[0, 1]}}
\usage{
branch_code_to_y(branch_code)
}
\arguments{
\item{branch_code}{a species' branching pattern,
as checked by \link{check_branch_code}, e.g. \code{A}, \code{AB}, \code{AABA}.}
}
\value{
a y coordinat in range \link{0, 1}
}
\description{
Convert a \code{branch_code} to a y coordinat in range \verb{[0, 1]}.
}
\details{
The y coordinat is the result of a division:
\itemize{
\item numerator: for a binary value of \code{b} (converted from the \code{branch_code},
where the \code{A} equals a zero, and the \code{B} equals a
one), the numerator in this division
equals \verb{1 + 2b}
\item denominator: for a \code{branch_code} of length \code{n},
the denominator equals \code{2 ^ n}.
}
}
\examples{
branch_code_to_y("A")
branch_code_to_y("AA")
branch_code_to_y("AB")
}
\author{
Richèl J.C. Bilderbeek
}
