% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ctt.R
\name{calc_ctt}
\alias{calc_ctt}
\title{Calculates the delta colonisation through time (CTT) statistic between the
ideal and empirical data sets simulated for each replicate}
\usage{
calc_ctt(daisie_data)
}
\arguments{
\item{daisie_data}{List containing data of DAISIE simulation with mainland
dynamics. Output from \code{sim_island_with_mainland}}
}
\value{
A vector of numerics
}
\description{
Calculates the delta colonisation through time (CTT) statistic between the
ideal and empirical data sets simulated for each replicate
}
\author{
Joshua W. Lambert
}
