% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ctt.R
\name{calc_ctt}
\alias{calc_ctt}
\title{Calculates the delta colonisation through time (CTT) statistic between the
ideal and empirical data sets simulated for each replicate}
\usage{
calc_ctt(daisie_data)
}
\arguments{
\item{daisie_data}{\link{list} containing data of of a DAISIE simulation
with mainland dynamics,
as produced by \link{sim_island_with_mainland}
and checked by \link{check_daisie_data}

A \code{daisie_data} has two elements:
\itemize{
\item \code{ideal_islands}: the ideal data set produced in a simulation,
see \link{check_ideal_islands} for more details
\item \code{empirical_islands}: the empirical data sets produced in a simulation
see \link{check_empirical_islands} for more details
}}
}
\value{
A vector of numerics
}
\description{
Calculates the delta colonisation through time (CTT) statistic between the
ideal and empirical data sets simulated for each replicate
}
\author{
Joshua W. Lambert
}
