% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_endemic_percent.R
\name{calc_endemic_percent}
\alias{calc_endemic_percent}
\title{Calculates what percentage of island species are endemic.}
\usage{
calc_endemic_percent(daisie_mainland_data)
}
\value{
A list of six numeric vectors
}
\description{
Calculates what percentage of island species are endemic.
}
\author{
Joshua W. Lambert
}
