% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error.R
\name{calc_error}
\alias{calc_error}
\title{Calculates the colonisation through time (CTT) statistic, percentage of max
age colonists, percentage of endemics, and differences in parameter estimates
from DAISIE maximum likelihood estimation on ideal and empirical data
simulated from sim_island_with_mainland}
\usage{
calc_error(daisie_data, ideal_ml, empirical_ml)
}
\arguments{
\item{daisie_data}{List containing data of DAISIE simulation with mainland
dynamics. Output from \code{sim_island_with_mainland}.

A list. The highest level of the list has two elements called
\code{ideal_islands} and \code{empirical_islands} which corresponds to
the ideal and empirical data sets produce in each simulation. Within each
\code{ideal_islands} and \code{empirical_islands} each element is an
individual replicate. The first element of each replicate is composed of
island information containing:
\describe{
\item{\verb{$island_age}}{A numeric with the island age.}
\item{\verb{$not_present}}{the number of mainland lineages that are not
present on the island.}
}
The subsequent elements of the list pertaining to each replicate contain
information on a single colonist lineage on the island and have 3 components:
\describe{
\item{\verb{$branching_times}}{island age and stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge and
Endemic anagenetic species.
For cladogenetic species these should
be island age and branching times of the radiation including the
stem age of the radiation.}
\item{\verb{$stac}}{An integer ranging from 1 to 6
indicating the status of the colonist:
\enumerate{
\item Non_endemic_MaxAge
\item Endemic
\item Endemic&Non_Endemic
\item Non_endemic_MaxAge
\item Endemic_singleton_MaxAge
\item Endemic_clade_MaxAge
}}
\item{\code{missing_species}}{number of island species that were
not sampled for particular clade (only applicable for endemic clades)}
}}

\item{ideal_ml}{List containing maximum likelihood estimates from DAISIE
fitted to ideal data produced from \code{sim_island_with_mainland}. Output
from \code{DAISIE::DAISIE_ML_CS}}

\item{empirical_ml}{List containing maximum likelihood estimates from DAISIE
fitted to empirical data produced from \code{sim_island_with_mainland}.
Output from \code{DAISIE::DAISIE_ML_CS}}
}
\value{
List of error metrics
}
\description{
Calculates the colonisation through time (CTT) statistic, percentage of max
age colonists, percentage of endemics, and differences in parameter estimates
from DAISIE maximum likelihood estimation on ideal and empirical data
simulated from sim_island_with_mainland
}
\author{
Joshua W. Lambert
}
