% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_island_endemics.R
\name{calc_island_endemics}
\alias{calc_island_endemics}
\title{Calculates the number of endemic and non-endemics species on a single island
replicate}
\usage{
calc_island_endemics(island)
}
\arguments{
\item{island}{A single island replicate, can be either ideal or empirical
data}
}
\value{
List of two numerics
}
\description{
Calculates the number of endemic and non-endemics species on a single island
replicate
}
\author{
Joshua W. Lambert
}
