% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_island_endemics.R
\name{calc_island_endemics}
\alias{calc_island_endemics}
\title{Calculates the number of endemic and non-endemics species on a single island
replicate}
\usage{
calc_island_endemics(island)
}
\arguments{
\item{island}{The history of a a single island,
as can be checked by \link{check_island}.

An \code{island} is \link{list} with two elements:
\itemize{
\item \code{ideal_island}: the island history based on ideal data,
as can be checked by \link{check_ideal_island}
\item \code{empirical_island}: the island history based on empirical data
as can be checked by \link{check_empirical_island}
For recolonising lineages, there is an extra element,
\code{all_colonisations} per list element.
It is comprised of \verb{$event_times} and \verb{$species_type}:
\describe{
\item{\verb{$event_times}}{ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times. Each vector pertains to one
colonising lineage.}
\item{\verb{$species_type}}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}
}}
}
\value{
List of two numerics
}
\description{
Calculates the number of endemic and non-endemics species on a single island
replicate
}
\author{
Joshua W. Lambert
}
