% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_island_endemics.R
\name{calc_island_endemics}
\alias{calc_island_endemics}
\title{Calculates the number of endemic and non-endemics species on a single island
replicate}
\usage{
calc_island_endemics(island)
}
\arguments{
\item{island}{The history of a a single island.
A \link{list} with two elements:
\itemize{
\item \code{ideal_island}: the island history based on ideal data
\item \code{empirical_island}: the island history based on empirical data
}

Each of these is a \link{list} with the following elements:
\itemize{
\item \verb{$branching_times}: island age and stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge and
Endemic anagenetic species.

For cladogenetic species these should
be island age and branching times of the radiation including the
stem age of the radiation.
\item \verb{$stac}: An integer ranging from 1 to 6
indicating the status of the colonist:
\enumerate{
\item Non_endemic_MaxAge
\item Endemic
\item Endemic&Non_Endemic
\item Non_endemic
\item Endemic_singleton_MaxAge
\item Endemic_clade_MaxAge
}
\item \verb{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)
}

For recolonising lineages, there is an extra element,
\code{all_colonisations} per list element.
It is comprised of \verb{$event_times} and \verb{$species_type}:
\describe{
\item{\verb{$event_times}}{ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times. Each vector pertains to one
colonising lineage.}
\item{\verb{$species_type}}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}}
}
\value{
List of two numerics
}
\description{
Calculates the number of endemic and non-endemics species on a single island
replicate
}
\author{
Joshua W. Lambert
}
