% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_island_endemics.R
\name{calc_island_endemics}
\alias{calc_island_endemics}
\title{Calculates the number of endemic and non-endemics species on a single island
replicate}
\usage{
calc_island_endemics(island)
}
\arguments{
\item{island}{The history of a a single island,
as can be checked by \link{check_island}.

An \code{island} is \link{list} with two elements:
\itemize{
\item \code{ideal_island}: the island history based on ideal data,
as can be checked by \link{check_ideal_island}
\item \code{empirical_island}: the island history based on empirical data
as can be checked by \link{check_empirical_island}
}}
}
\value{
List of two numerics
}
\description{
Calculates the number of endemic and non-endemics species on a single island
replicate
}
\author{
Joshua W. Lambert
}
