% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_num_col.R
\name{calc_num_col}
\alias{calc_num_col}
\title{Calculates summary metrics from a simulation}
\usage{
calc_num_col(multi_daisie_data)
}
\arguments{
\item{multi_daisie_data}{A list of \code{daisie_data} elements}
}
\value{
Numeric vector number of colonisations for each island replicate
}
\description{
Calculates summary metrics from a simulation
}
\author{
Joshua W. Lambert
}
