% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_num_spec.R
\name{calc_num_spec}
\alias{calc_num_spec}
\title{Calculates summary metrics from a simulation}
\usage{
calc_num_spec(multi_daisie_data)
}
\arguments{
\item{multi_daisie_data}{A list of \code{daisie_data} elements}
}
\value{
Numeric vector number of species for each island replicate
}
\description{
Calculates summary metrics from a simulation
}
\author{
Joshua W. Lambert
}
