% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rates.R
\name{calc_rates}
\alias{calc_rates}
\title{Calculate rates for island processes}
\usage{
calc_rates(
  timeval,
  totaltime,
  gam,
  laa,
  lac,
  mu,
  k,
  num_spec,
  num_immigrants,
  mainland_n,
  island_spec = NULL
)
}
\arguments{
\item{timeval}{Numeric defining current time of simulation.}

\item{totaltime}{Numeric defining the length of the simulation in time units.}

\item{gam}{Numeric with the per capita immigration rate.}

\item{laa}{Numeric with the per capita anagenesis rate.}

\item{lac}{Numeric with the per capita cladogenesis rate.}

\item{mu}{Numeric with the per capita extinction rate.}

\item{k}{Numeric with carrying capacity.}

\item{num_spec}{Numeric with the current number of species.}

\item{num_immigrants}{Numeric with the current number of non-endemic
species.}

\item{mainland_n}{Numeric stating the number of mainland species that
can colonize the island.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}
}
\value{
Named list with rates.
}
\description{
Internal function that updates the all the rates at time t.
}
\keyword{internal}
