% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sim_metrics.R
\name{calc_sim_metrics}
\alias{calc_sim_metrics}
\title{Calculates the summary metrics for the entire parameter space}
\usage{
calc_sim_metrics(analysis_results, output_file_path)
}
\arguments{
\item{analysis_results}{A list of analysis result elements which can be
checked with \code{check_analysis_result}}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}
}
\value{
List of simulation metrics
}
\description{
Calculates the summary metrics for the entire parameter space
}
\author{
Joshua W. Lambert
}
