% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all_colonisations.R
\name{check_all_colonisations}
\alias{check_all_colonisations}
\title{Check if a \code{all_colonisations} is valid.}
\usage{
check_all_colonisations(all_colonisations)
}
\arguments{
\item{all_colonisations}{a list of one or more \code{colonisations},
as can be checked by \link{check_all_colonisations}.
See \link{check_colonisations} for details on the \code{colonisations}}
}
\value{
nothing
}
\description{
Check if a \code{all_colonisations} is valid.
Will \link{stop} if not
}
\details{
\code{all_colonisations} is a list of one or more colonisations,
where each element is checked by \link{check_colonisations}.
See \link{check_colonisations} for more details.

A valid \code{all_colonisations} ...
\itemize{
\item is a list of one or more \code{colonisations},
where each \code{colonisations} is checked by \link{check_colonisations}
}
}
\examples{
set.seed(
  2,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = 20
)
island <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 12, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
all_colonisations <- island$ideal_island[[2]]$all_colonisations
check_all_colonisations(all_colonisations)

}
\author{
Richèl J.C. Bilderbeek
}
