% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_colonisations.R
\name{check_colonisations}
\alias{check_colonisations}
\title{Check if an \code{colonisations} is valid.}
\usage{
check_colonisations(colonisations)
}
\arguments{
\item{colonisations}{a \code{colonisations},
as can be checked by \link{check_colonisations}.}
}
\value{
nothing
}
\description{
Check if an \code{colonisations} is valid.
Will \link{stop} if not.
}
\details{
A \code{colonisations} contains:
\itemize{
\item a \code{species_type}, which is the origin of the clade,
which can be anagenetic, cladogenetic or immigration.
See \link{check_species_type} for details
\item a \code{event_times}, which holds events such as the island age,
stem age of the species and -for cladogenetic species only-,
the branching times of the radiation.
see \link{check_event_times}
}
}
\examples{
set.seed(
  2,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = 20
)
island <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 12, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
colonisations <- island$ideal_island[[2]]$all_colonisations[[1]]
check_colonisations(colonisations)
}
\author{
Richèl J.C. Bilderbeek
}
