% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_colonisations.R
\name{check_colonisations}
\alias{check_colonisations}
\title{Check if an \code{colonisations} is valid.}
\usage{
check_colonisations(colonisations)
}
\value{
nothing
}
\description{
Check if an \code{colonisations} is valid.
Will \link{stop} if not.
}
\examples{
set.seed(
  2,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
mainland_clade <- create_test_mainland_clade(
  mainland_scenario = 20
)
island <- sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 12, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
colonisations <- island$ideal_island[[2]]$all_colonisations[[1]]
check_colonisations(colonisations)
}
\author{
Richèl J.C. Bilderbeek
}
