% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_daisie_data.R
\name{check_daisie_data}
\alias{check_daisie_data}
\title{Check if a \code{daisie_data} is valid.
Will \link{stop} if not.}
\usage{
check_daisie_data(daisie_data)
}
\arguments{
\item{daisie_data}{List containing data of DAISIE simulation with mainland
dynamics. Output from \code{sim_island_with_mainland}.

A list. The highest level of the list has two elements called
\code{ideal_islands} and \code{empirical_islands} which corresponds to
the ideal and empirical data sets produce in each simulation. Within each
\code{ideal_islands} and \code{empirical_islands} each element is an
individual replicate. The first element of each replicate is composed of
island information containing:
\describe{
\item{\verb{$island_age}}{A numeric with the island age.}
\item{\verb{$not_present}}{the number of mainland lineages that are not
present on the island.}
}
The subsequent elements of the list pertaining to each replicate contain
information on a single colonist lineage on the island and have 3 components:
\describe{
\item{\verb{$branching_times}}{island age and stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge and
Endemic anagenetic species.
For cladogenetic species these should
be island age and branching times of the radiation including the
stem age of the radiation.}
\item{\verb{$stac}}{An integer ranging from 1 to 6
indicating the status of the colonist:
\enumerate{
\item Non_endemic_MaxAge
\item Endemic
\item Endemic&Non_Endemic
\item Non_endemic_MaxAge
\item Endemic_singleton_MaxAge
\item Endemic_clade_MaxAge
}}
\item{\code{missing_species}}{number of island species that were
not sampled for particular clade (only applicable for endemic clades)}
}}
}
\value{
nothing
}
\description{
Check if a \code{daisie_data} is valid.
Will \link{stop} if not.
}
\examples{
daisie_data <- sim_island_with_mainland(
  total_time = 1,
  m = 100,
  island_pars = c(1, 1, 50, 0.1, 1),
  mainland_ex = 0.5,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete",
  replicates = 1,
  verbose = FALSE
)
check_daisie_datalist(daisie_data)

}
\author{
Richèl J.C. Bilderbeek
}
