% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_island.R
\name{check_island}
\alias{check_island}
\title{Check if an \code{island} is valid.}
\usage{
check_island(island)
}
\arguments{
\item{island}{The history of a a single island,
as can be checked by \link{check_island}.

An \code{island} is \link{list} with two elements:
\itemize{
\item \code{ideal_island}: the island history based on ideal data,
as can be checked by \link{check_ideal_island}
\item \code{empirical_island}: the island history based on empirical data
as can be checked by \link{check_empirical_island}
}}
}
\value{
nothing
}
\description{
Check if an \code{island} is valid.
Will \link{stop} if not.
}
\examples{
set.seed(
  2,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = 2)
island <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 1, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")

check_island(island)
}
\author{
Richèl J.C. Bilderbeek
}
