% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_species.R
\name{check_missing_species}
\alias{check_missing_species}
\title{Check if a \code{missing_species} is valid.}
\usage{
check_missing_species(missing_species)
}
\value{
nothing
}
\description{
Check if a \code{missing_species} is valid.
Will \link{stop} if not
}
\details{
A valid \code{missing_species} ...
\itemize{
\item is numeric
\item has one element
\item the element has a whole-number value in range \verb{[0, Inf>}
}
}
\examples{
check_missing_species(1)
}
\seealso{
Use \link{missing_species_to_str} to convert a \code{missing_species} to a string.
}
\author{
Richèl J.C. Bilderbeek
}
