% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clado_event.R
\name{clado_event}
\alias{clado_event}
\title{Updates the island_spec and max_spec_id when a cladogenesis event occurs}
\usage{
clado_event(timeval, island_spec, max_spec_id)
}
\arguments{
\item{timeval}{Numeric defining current time of simulation.}

\item{island_spec}{Matrix with current state of simulation containing number
of species, see \code{\link[=create_test_island_spec]{create_test_island_spec()}}}

\item{max_spec_id}{Current species IDs.}
}
\value{
A two element list
}
\description{
Updates the island_spec and max_spec_id when a cladogenesis event occurs
}
\author{
Joshua W. Lambert
}
