% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{collect_empirical_island_stacs}
\alias{collect_empirical_island_stacs}
\title{Collect the \code{stac}s of an \code{empirical_island}}
\usage{
collect_empirical_island_stacs(empirical_island)
}
\arguments{
\item{empirical_island}{a simulated island with imperfect information,
with one \code{empirical_island_clade} per simulated island clade.

An \code{empirical_island} is created by \link{sim_island}
(together with \code{ideal_island}, which has perfect information),
as can be checked by \link{check_empirical_island}.}
}
\value{
a numeric vector of \code{stac}s
}
\description{
Collect the \code{stac}s of an \code{empirical_island}
}
\author{
Richèl J.C. Bilderbeek
}
