% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{collect_ideal_island_stacs}
\alias{collect_ideal_island_stacs}
\title{Collect the \code{stac}s of an \code{ideal_island}}
\usage{
collect_ideal_island_stacs(ideal_island)
}
\arguments{
\item{ideal_island}{a simulated island with perfect information,
which is a \link{list} of one or more \link{ideal_island_clade}s.

An \code{ideal_island} is created by \link{sim_island} (together with \code{empirical_island},
which has imperfect information),
as can be checked by \link{check_ideal_island}}
}
\value{
a numeric vector of \code{stac}s
}
\description{
Collect the \code{stac}s of an \code{ideal_island}
}
\author{
Richèl J.C. Bilderbeek
}
