% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colonist_species_type_to_str.R
\name{colonist_species_type_to_str}
\alias{colonist_species_type_to_str}
\title{Convert a \code{colonist_species_type} to a string}
\usage{
colonist_species_type_to_str(colonist_species_type)
}
\value{
the \code{colonist_species_type} as a string
}
\description{
Convert a \code{colonist_species_type} to a string:
}
\details{
\tabular{ll}{
   Value \tab String \cr
   \code{A} \tab \verb{Anagenetic recolonist} \cr
   \code{C} \tab \verb{Cladogenetic recolonist} \cr
}
}
\examples{
DAISIEmainland:::colonist_species_type_to_str("A")
DAISIEmainland:::colonist_species_type_to_str("C")
}
\author{
Richèl J.C. Bilderbeek
}
\keyword{internal}
