% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{create_labels}
\alias{create_labels}
\title{Creates labels for plots by calling \code{scientific}.}
\usage{
create_labels(signif)
}
\value{
A function that takes a vector
}
\description{
Creates labels for plots by calling \code{scientific}.
}
