% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{create_labels}
\alias{create_labels}
\title{Creates labels for plots by calling \code{scientific}.}
\usage{
create_labels(signif, scientific)
}
\arguments{
\item{signif}{A numeric specifying how many significant figures the axes
labels have when plotting}

\item{scientific}{A boolean determining whether the axis labels will be
converted to scientific notation}
}
\value{
A function that takes a numeric vector
}
\description{
Creates labels for plots by calling \code{scientific}.
}
