% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_non_empty_island.R
\name{create_non_empty_island}
\alias{create_non_empty_island}
\title{Creates an ideal and empirical data set for a single non empty island}
\usage{
create_non_empty_island(
  total_time,
  island_spec,
  mainland_clade,
  mainland_sample_prob
)
}
\arguments{
\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}

\item{mainland_clade}{Data frame with state of mainland. see
\code{\link{create_test_mainland_clade}}.}

\item{mainland_sample_prob}{Numeric between zero and one determining the
probability of an extant mainland species being sampled.}
}
\value{
List with the island information, composed of
branching times of extant species, status of species on
the island and number of missing species.
}
\description{
Creates an ideal and empirical data set for a single non empty island
}
\author{
Joshua W. Lambert
}
