% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_island_spec.R
\name{create_test_island_spec}
\alias{create_test_island_spec}
\title{Creates an island community for testing purposes}
\usage{
create_test_island_spec(island_scenario)
}
\value{
data frame with island community information
}
\description{
Creates an island community for testing purposes
}
\keyword{internal}
