% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_data_to_tables.R
\name{daisie_data_colonist_info_to_colonisation_times_table}
\alias{daisie_data_colonist_info_to_colonisation_times_table}
\title{Internal function}
\usage{
daisie_data_colonist_info_to_colonisation_times_table(
  daisie_data_colonist_info
)
}
\arguments{
\item{daisie_data_colonist_info}{an element of a \code{daisie_data},
that is not the first element (the first element is of type
\code{daisie_data_header}).}
}
\value{
a table with as much rows as colonisation times
}
\description{
Convert the non-header part of a \code{daisie_data} into a tabular format,
containing, for each colonist, the colonisation time.
}
\details{
Due to the coding of DAISIE, this is not straightforward:
\itemize{
\item If there are no colonisations, an empty table is produced
\item If there is one colonisation, the second \code{event_times} are used
(the first \code{event_times} is the island age)
\item If there are two or more colonisations, the second \code{event_times}
in each \code{all_colonisations} element is used
(the first \code{event_times} is the island age)
}
}
\seealso{
use \link{daisie_data_colonist_info_to_braching_times_table}
to obtain the branching times of the colonists
}
\author{
Richèl J.C. Bilderbeek
}
