% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_data_to_tables.R
\name{daisie_data_to_tables}
\alias{daisie_data_to_tables}
\title{Convert an \code{daisie_data} to a list of tables}
\usage{
daisie_data_to_tables(daisie_data)
}
\arguments{
\item{daisie_data}{List containing data of DAISIE simulation with mainland
dynamics. Output from `sim_island_with_mainland``}
}
\value{
a \link{list} with elements:
\itemize{
\item \code{speciations}: a table with the speciation events
\item \code{colonisations}: a table with colonisations
}
}
\description{
Convert an \code{daisie_data} to a list of tables
}
\author{
Richèl J.C. Bilderbeek
}
