% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Documentation for function arguments in the DAISIEmainland package}
\usage{
default_params_doc(
  all_colonisations,
  branch_code,
  branching_times,
  colonisations,
  event_times,
  missing_species,
  stac,
  timeval,
  total_time,
  gam,
  laa,
  lac,
  mu,
  k,
  num_spec,
  num_immigrants,
  mainland_n,
  island_spec,
  mainland_clade,
  mainland_spec,
  mainland_sample_prob,
  mainland_sample_type,
  island_replicates,
  m,
  verbose,
  rates,
  island_pars,
  mainland_ex,
  replicates,
  possible_event,
  max_spec_id,
  mainland_scenario,
  anc_branch_t_bp,
  subset_island,
  daisie_datalist,
  daisie_data,
  ideal_daisie_data,
  empirical_daisie_data,
  ideal_island,
  ideal_island_clade,
  empirical_island,
  empirical_island_clade,
  ideal_or_empirical_island,
  species_type,
  ideal_ml,
  empirical_ml,
  sim_params,
  param_set,
  data_folder_path,
  output_file_path,
  param_space,
  parameter,
  plotting_data,
  breaks,
  num_breaks,
  signif,
  mainland,
  island
)
}
\arguments{
\item{all_colonisations}{a list of one or more \code{colonisations},
see \link{check_colonisations} for details.
\code{all_colonisations} can be checked by \link{check_all_colonisations}.

\code{all_colonisations} is an element of
\itemize{
\item \code{ideal_island_clade} (see \link{check_ideal_island_clade})
\item \code{emprical_island_clade} (see \link{check_empirical_island_clade})
}}

\item{branch_code}{a species' branching pattern,
as checked by \link{check_branch_code}, e.g. \code{A}, \code{AB}, \code{AABA}.}

\item{branching_times}{island age and stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge and
Endemic anagenetic species.

For cladogenetic species these should
be island age and branching times of the radiation including the
stem age of the radiation.

See \link{check_branching_times} for a more detailed description.}

\item{colonisations}{a \code{colonisations}, which is a \link{list} with elements:
\itemize{
\item \code{event_times}: the times that events (i.e. island age, immigration,
speciation) occurred (see \link{check_event_times})
\item \code{species_type}:  the origin of a clade (i.e. anagenetic, cladogenetic
or immigrant) (see \link{check_species_type})
}

An \code{all_colonisations} (see \link{check_all_colonisations}) is a \link{list}
of \code{colonisations}.}

\item{event_times}{the island age, stem age of species,
and -for cladogenetic species- speciation times,
as checked by \link{check_event_times}}

\item{missing_species}{the number of island species that were not
sampled for particular clade (only applicable for endemic clades),
as can be checked by \link{check_missing_species}}

\item{stac}{the status of the colonist,
see \link{stac_to_str} for values and meanings}

\item{timeval}{Numeric defining current time of simulation.}

\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{gam}{Numeric with the per capita immigration rate.}

\item{laa}{Numeric with the per capita anagenesis rate.}

\item{lac}{Numeric with the per capita cladogenesis rate.}

\item{mu}{Numeric with the per capita extinction rate.}

\item{k}{Numeric with carrying capacity.}

\item{num_spec}{Numeric with the current number of species.}

\item{num_immigrants}{Numeric with the current number of non-endemic
species.}

\item{mainland_n}{Numeric stating the number of mainland species that
can colonize the island.}

\item{island_spec}{Matrix with current state of simulation containing number
of species, see \code{\link[=create_test_island_spec]{create_test_island_spec()}}}

\item{mainland_clade}{Data frame with state of mainland, see
\code{\link[=create_test_mainland_clade]{create_test_mainland_clade()}}}

\item{mainland_spec}{Numeric focal species on the mainland}

\item{mainland_sample_prob}{Numeric between zero and one determining the
probability of an extant mainland species being sampled.}

\item{mainland_sample_type}{String either "unsampled" for unsampled known
species on the mainland, "undiscovered" for undiscovered species on the
mainland, or "complete" if the mainland sampling probability is one.}

\item{island_replicates}{a \link{list} of that has as many elements as
replicates. Each element is a \link{list} of \code{islands} (as checked
by \link{check_island}).
Use \link{check_island_replicates} to check a \code{island_replicates}
being correct}

\item{m}{Numeric defining the size of mainland pool.}

\item{verbose}{Logical, determining if progress output should be printed
during the simulation.}

\item{rates}{named list of numeric rates as returned by \link{calc_rates}.}

\item{island_pars}{A numeric vector containing the parameters for the island:
\itemize{
\item{\code{island_pars[1]}: lambda^c (cladogenesis rate)}
\item{\code{island_pars[2]}: mu (extinction rate)}
\item{\code{island_pars[3]}: K (carrying capacity), set K=Inf for
diversity independence.}
\item{\code{island_pars[4]}: gamma (immigration rate)}
\item{\code{island_pars[5]}: lambda^a (anagenesis rate)}
}}

\item{mainland_ex}{Numeric parameter for mainland extinction rate.}

\item{replicates}{Integer specifying number of island replicates to be
simulated.}

\item{possible_event}{Numeric defining what event will happen.}

\item{max_spec_id}{Current species IDs.}

\item{mainland_scenario}{Integer determining which mainland clade scenario
is produced}

\item{anc_branch_t_bp}{Numeric ancestral branching time before the present
of the immigrant species ancestor and its extant relative on the mainland}

\item{subset_island}{Data frame of island species that originated from a
single colonisation event}

\item{daisie_datalist}{a \link[DAISIE]{DAISIE} \code{datalist}, as can be
checked by \link{check_daisie_datalist}. To quote the DAISE doc:

A \code{datalist} is an object containing information on colonisation and
branching times. This object can be generated using
\link[DAISIE]{DAISIE_dataprep},
which converts a user-specified data table into a \code{datalist},
but the object can of course also be entered directly.

A \code{datalist} is a \link{list} with the following elements:

The first element of the list has two or three components:
\itemize{
\item \code{$island_age} - the island age
}

Then, depending on whether a distinction between types is
made, we have:
\itemize{
\item \code{$not_present} - the number of mainland lineages
that are not present on the island \cr
}

or:
\itemize{
\item \code{$not_present_type1} - the number of mainland lineages
of type 1 that are not present on the island
\item \code{$not_present_type2} - the number of mainland lineages of
type 2 that are not present on the island
}

The remaining elements of
the \code{datalist} each contains information on a single colonist lineage on the
island and has 5 components:
\itemize{
\item \code{$colonist_name} - the name of the species or clade that
colonized the island
\item \code{$branching_times} - island age followed by stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge
species and Endemic species with no close relatives
on the island. For endemic clades with more than one species on the island
(cladogenetic clades/ radiations) these should be island age followed by the
branching times of the island clade
including the stem age of the clade
\item \code{$stac} - the status of the colonist, see \link{stac_to_str}
for values and descriptions
\item \code{$missing_species} - number of island species that were not
sampled for particular clade (only applicable for endemic clades)
\item \code{$type1or2} - whether the colonist belongs to type 1 or type 2
}}

\item{daisie_data}{\link{list} containing data of of a DAISIE simulation
with mainland dynamics,
as produced by \link{sim_island_with_mainland}
and checked by \link{check_daisie_data}

A \code{daisie_data} has two elements:
\itemize{
\item \code{ideal_islands}: the ideal data set produced in a simulation,
see \link{check_ideal_daisie_data} for more details
\item \code{empirical_islands}: the empirical data sets produced in a simulation
see \link{check_empirical_daisie_data} for more details
}}

\item{ideal_daisie_data}{a \link{list},
of which each element contains the history
of a DAISIE simulation with mainland dynamics, with a length that
equals the number of replicates,
as can be checked by \link{check_ideal_daisie_data}

The simulated history is recorded perfectly, resulting in ideal data.
The list elements are of type \code{DAISIE::datalist},
as checked by \link{check_daisie_datalist}
and can be used by \link[DAISIE]{DAISIE} (e.g. \link[DAISIE]{DAISIE_ML_CS})}

\item{empirical_daisie_data}{a list, of which each element contains the
history of a DAISIE simulation with mainland dynamics, with a length that
equals the number of replicates,
as can be checked by \link{check_empirical_daisie_data}

The simulated history is recorded as an empiricist would,
resulting in imperfect data.
The list elements are of type \code{DAISIE::datalist},
as checked by \link{check_daisie_datalist}
and can be used by \link[DAISIE]{DAISIE} (e.g. \link[DAISIE]{DAISIE_ML_CS})}

\item{ideal_island}{a simulated island with perfect information,
which is a \link{list} of one or more \code{ideal_island_clade}s.

An \code{ideal_island} is created by \link{sim_island} (together with \code{empirical_island},
which has imperfect information),
as can be checked by \link{check_ideal_island}}

\item{ideal_island_clade}{A simulated island clade with perfect information,
which is a \link{list} of one or more \code{ideal_island_clade}s,
as can be checked by \link{check_ideal_island_clade}

\code{ideal_island_clade} is a list  containing 3 components:
\itemize{
\item \verb{$branching_times}: island age, species stem age and speciation times.
See \link{check_branching_times} for more details
\item \verb{$stac}: the status of the colonist,
see \link{stac_to_str} for values and meanings
\item \verb{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)
}

For recolonising lineages, there is an extra element,
\code{all_colonisations} per list element.
It is comprised of \verb{$event_times} and \verb{$species_type}:
\describe{
\item{\verb{$event_times}}{ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times. Each vector pertains to one
colonising lineage.}
\item{\verb{$species_type}}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}}

\item{empirical_island}{a simulated island with imperfect information,
with one \code{empirical_island_clade} per simulated island clade.

An \code{empirical_island} is created by \link{sim_island}
(together with \code{ideal_island}, which has perfect information),
as can be checked by \link{check_empirical_island}.}

\item{empirical_island_clade}{A simulated island clade with imperfect information,
as can be checked by \link{check_empirical_island_clade}.

An \code{empirical_island_clade} contains 3 components:
\itemize{
\item \verb{$branching_times}: island age, species stem age and speciation times.
See \link{check_branching_times} for more details
\item \verb{$stac}: the status of the colonist,
see \link{stac_to_str} for values and meanings
\item \verb{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)
}

For recolonising lineages, there is an extra element,
\code{all_colonisations} per list element.
It is comprised of \verb{$event_times} and \verb{$species_type}:
\describe{
\item{\verb{$event_times}}{ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times. Each vector pertains to one
colonising lineage.}
\item{\verb{$species_type}}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}}

\item{ideal_or_empirical_island}{the evolutionary history of
an island from ideal or empirical data,
as created by \link{sim_island}

\code{ideal_or_empirical_island} is a list  containing 3 components:
\itemize{
\item \verb{$branching_times}: island age, species stem age and speciation times.
See \link{check_branching_times} for more details
\item \verb{$stac}: the status of the colonist,
see \link{stac_to_str} for values and meanings
\item \verb{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)
}

For recolonising lineages, there is an extra element,
\code{all_colonisations} per list element.
It is comprised of \verb{$event_times} and \verb{$species_type}:
\describe{
\item{\verb{$event_times}}{ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times. Each vector pertains to one
colonising lineage.}
\item{\verb{$species_type}}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}}

\item{species_type}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}

\item{ideal_ml}{List containing maximum likelihood estimates from DAISIE
fitted to ideal data produced from \code{sim_island_with_mainland}. Output
from \code{DAISIE::DAISIE_ML_CS}}

\item{empirical_ml}{List containing maximum likelihood estimates from DAISIE
fitted to empirical data produced from \code{sim_island_with_mainland}.
Output from \code{DAISIE::DAISIE_ML_CS}}

\item{sim_params}{Vector of five numerics for the values of cladogenesis,
extinction, carrying capacity, immigration and anagenesis used to simulate
data}

\item{param_set}{Numeric for the index of the parameter set from the
parameter space}

\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{param_space}{String specifying which parameter space results to read.
Either "general", "mainland_ex", or "mainland_sample_prob"}

\item{parameter}{String specifying which parameter is plotted.
"mainland_ex" plots mainland extinction as the variable,
"mainland_sample_prob" plots mainland sampling probability as the variable
and "both" plots all the data}

\item{plotting_data}{Numeric vector of data points}

\item{breaks}{A vector of numerics}

\item{num_breaks}{A numeric specifying how many breaks are wanted on the
plot axes}

\item{signif}{A numeric specifying how many significant figures the axes
labels have when plotting}

\item{mainland}{the evolutionary history of the mainland species,
as created by \link{sim_mainland}.
Use \link{plot_mainland} to visualise the that evolutionary history.}

\item{island}{The history of a a single island,
as can be checked by \link{check_island}.

An \code{island} is \link{list} with two elements:
\itemize{
\item \code{ideal_island}: the island history based on ideal data,
as can be checked by \link{check_ideal_island}
\item \code{empirical_island}: the island history based on empirical data
as can be checked by \link{check_empirical_island}
}}
}
\value{
Nothing
}
\description{
Documentation for function arguments in the DAISIEmainland package
}
\author{
Joshua W. Lambert
}
