% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Documentation for function arguments in the DAISIEmainland package}
\usage{
default_params_doc(
  timeval,
  total_time,
  gam,
  laa,
  lac,
  mu,
  k,
  num_spec,
  num_immigrants,
  mainland_n,
  island_spec,
  mainland_clade,
  mainland_spec,
  mainland_sample_prob,
  island_replicates,
  m,
  verbose,
  rates,
  island_pars,
  mainland_ex,
  replicates,
  possible_event,
  max_spec_id,
  mainland_scenario,
  anc_branch_t_bp,
  subset_island,
  daisie_data,
  ideal_ml,
  empirical_ml,
  sim_params,
  param_set,
  xlim,
  data_folder_path,
  output_file_path,
  param_space,
  parameter
)
}
\arguments{
\item{timeval}{Numeric defining current time of simulation.}

\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{gam}{Numeric with the per capita immigration rate.}

\item{laa}{Numeric with the per capita anagenesis rate.}

\item{lac}{Numeric with the per capita cladogenesis rate.}

\item{mu}{Numeric with the per capita extinction rate.}

\item{k}{Numeric with carrying capacity.}

\item{num_spec}{Numeric with the current number of species.}

\item{num_immigrants}{Numeric with the current number of non-endemic
species.}

\item{mainland_n}{Numeric stating the number of mainland species that
can colonize the island.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}

\item{mainland_clade}{Data frame with state of mainland. see
\code{\link[=create_test_mainland_clade]{create_test_mainland_clade()}}.}

\item{mainland_spec}{Numeric focal species on the mainland}

\item{mainland_sample_prob}{Numeric between zero and one determining the
probability of an extant mainland species being sampled.}

\item{island_replicates}{List that has as many elements as replicates. Each
element must be a list with the elements \code{island_age} and
\code{not_present}. ##### LOOK INTO THIS}

\item{m}{Numeric defining the size of mainland pool.}

\item{verbose}{Logical, determining if progress output should be printed
during the simulation.}

\item{rates}{named list of numeric rates as returned by
\code{\link[=update_rates]{update_rates()}}.}

\item{island_pars}{A numeric vector containing the parameters for the island:
\itemize{
\item{\code{island_pars[1]}: lambda^c (cladogenesis rate)}
\item{\code{island_pars[2]}: mu (extinction rate)}
\item{\code{island_pars[3]}: K (carrying capacity), set K=Inf for
diversity independence.}
\item{\code{island_pars[4]}: gamma (immigration rate)}
\item{\code{island_pars[5]}: lambda^a (anagenesis rate)}
}}

\item{mainland_ex}{Numeric parameter for mainland extinction rate.}

\item{replicates}{Integer specifying number of island replicates to be
simulated.}

\item{possible_event}{Numeric defining what event will happen.}

\item{max_spec_id}{Current species IDs.}

\item{mainland_scenario}{Integer determining which mainland clade scenario
is produced}

\item{anc_branch_t_bp}{Numeric ancestral branching time before the present
of the immigrant species ancestor and its extant relative on the mainland}

\item{subset_island}{Data frame of island species that originated from a
single colonisation event}

\item{daisie_data}{List containing data of DAISIE simulation with mainland
dynamics. Output from `sim_island_with_mainland``}

\item{ideal_ml}{List containing maximum likelihood estimates from DAISIE
fitted to ideal data produced from \code{sim_island_with_mainland}. Output
from \code{DAISIE::DAISIE_ML_CS}}

\item{empirical_ml}{List containing maximum likelihood estimates from DAISIE
fitted to empirical data produced from \code{sim_island_with_mainland}.
Output from \code{DAISIE::DAISIE_ML_CS}}

\item{sim_params}{Vector of five numerics for the values of cladogenesis,
extinction, carrying capacity, immigration and anagenesis used to simulate
data}

\item{param_set}{Numeric for the index of the parameter set from the
parameter space}

\item{xlim}{Boolean determining whether the x-axis should be bounded on the
plot}

\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{param_space}{String specifying which parameter space results to read.
Either "general", "mainland_ex", or "mainland_sample_prob"}

\item{parameter}{String specifying which parameter is plotted.
"mainland_ex" plots mainland extinction as the variable,
"mainland_sample_prob" plots mainland sampling probability as the variable
and "both" plots all the data}
}
\value{
Nothing
}
\description{
Documentation for function arguments in the DAISIEmainland package
}
\author{
Joshua W. Lambert
}
