% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_to_tables.R
\name{empirical_island_to_tables}
\alias{empirical_island_to_tables}
\title{Convert an \code{empirical_island} to a list of tables}
\usage{
empirical_island_to_tables(empirical_island)
}
\arguments{
\item{empirical_island}{a simulated island with imperfect information,
as created by \link{sim_island} (together with \code{ideal_island},
which has perfect information).

\code{empirical_island} is a list  containing 3 components:
\itemize{
\item \verb{$branching_times}: island age and stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge and
Endemic anagenetic species.

For cladogenetic species these should
be island age and branching times of the radiation including the
stem age of the radiation.
\item \verb{$stac}: An integer ranging from 1 to 6
indicating the status of the colonist:
\enumerate{
\item Non_endemic_MaxAge
\item Endemic
\item Endemic&Non_Endemic
\item Non_endemic
\item Endemic_singleton_MaxAge
\item Endemic_clade_MaxAge
}
\item \verb{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)
}}
}
\value{
a \link{list} with elements:
\itemize{
\item \code{t}: a table
\item \code{all_colonisations}: a list with all colonizations
}
}
\description{
Convert an \code{empirical_island} to a list of tables
}
