% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_to_tables.R
\name{empirical_island_to_tables}
\alias{empirical_island_to_tables}
\title{Convert an \code{empirical_island} to a list of tables}
\usage{
empirical_island_to_tables(empirical_island)
}
\arguments{
\item{empirical_island}{a simulated island with imperfect information,
as created by \link{sim_island} (together with \code{ideal_island},
which has perfect information).

\code{empirical_island} is a list  containing 3 components:
\itemize{
\item \verb{$branching_times}: island age and stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge and
Endemic anagenetic species.

For cladogenetic species these should
be island age and branching times of the radiation including the
stem age of the radiation.
\item \verb{$stac}: An integer ranging from 1 to 6
indicating the status of the colonist:
\enumerate{
\item Non_endemic_MaxAge
\item Endemic
\item Endemic&Non_Endemic
\item Non_endemic
\item Endemic_singleton_MaxAge
\item Endemic_clade_MaxAge
}
\item \verb{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)
}

For recolonising lineages, there is an extra element,
\code{all_colonisations} per list element.
It is comprised of \verb{$event_times} and \verb{$species_type}:
\describe{
\item{\verb{$event_times}}{ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times. Each vector pertains to one
colonising lineage.}
\item{\verb{$species_type}}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}}
}
\value{
a \link{list} with elements:
\itemize{
\item \code{speciations}: a table with the speciation events
\item \code{colonisations}: a table with colonisations
}
}
\description{
Convert an \code{empirical_island} to a list of tables
}
