% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_utils.R
\name{event_to_str}
\alias{event_to_str}
\title{Convert an event to a string}
\usage{
event_to_str(possible_event)
}
\arguments{
\item{possible_event}{Numeric defining what event will happen.}
}
\value{
the event as a string
}
\description{
Convert an event to a string:
}
\details{
\tabular{ll}{
   Value \tab String \cr
   1 \tab \code{immigration} \cr
   2 \tab \code{extinction} \cr
   3 \tab \code{anagenesis} \cr
   4 \tab \code{cladogenesis} \cr
}
}
\examples{
DAISIEmainland:::event_to_str(1)
DAISIEmainland:::event_to_str(2)
DAISIEmainland:::event_to_str(3)
DAISIEmainland:::event_to_str(4)
}
\seealso{
use \link{str_to_event} to convent a string to an \code{event}
}
\author{
Richèl J.C. Bilderbeek
}
\keyword{internal}
