% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_event.R
\name{ext_event}
\alias{ext_event}
\title{Updates the island_tbl when an extinction event occurs}
\usage{
ext_event(island_tbl)
}
\arguments{
\item{island_tbl}{Data frame with current state of simulation containing
number of species, see \code{\link[=create_test_island_tbl]{create_test_island_tbl()}}}
}
\value{
A data frame
}
\description{
Updates the island_tbl when an extinction event occurs
}
\author{
Joshua W. Lambert
}
