% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_event.R
\name{ext_event}
\alias{ext_event}
\title{Updates the island_spec when an extinction event occurs}
\usage{
ext_event(island_spec)
}
\arguments{
\item{island_spec}{Matrix with current state of simulation containing number
of species.}
}
\value{
A data frame
}
\description{
Updates the island_spec when an extinction event occurs
}
