% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_multi_daisie_data.R
\name{group_multi_daisie_data}
\alias{group_multi_daisie_data}
\title{Title}
\usage{
group_multi_daisie_data(multi_daisie_data, total_time, m)
}
\arguments{
\item{m}{}
}
\value{

}
\description{
Title
}
