% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daisie_data_to_tables.R
\name{ideal_daisie_data_to_tables}
\alias{ideal_daisie_data_to_tables}
\title{Convert an \code{ideal_island} to a list of tables}
\usage{
ideal_daisie_data_to_tables(ideal_daisie_data)
}
\arguments{
\item{ideal_daisie_data}{a \link{list},
of which each element contains the history
of a DAISIE simulation with mainland dynamics, with a length that
equals the number of replicates.
The simulated history is recorded perfectly, resulting in ideal data.
The list elements are of type \code{DAISIE::datalist},
as checked by \link{check_daisie_datalist}
and can be used by \link[DAISIE]{DAISIE} (e.g. \link[DAISIE]{DAISIE_ML_CS})}
}
\value{
a \link{list} with elements:
\itemize{
\item \code{speciations}: a table with the speciation events
\item \code{colonisations}: a table with colonisations
}
}
\description{
Convert an \code{ideal_island} to a list of tables
}
