% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_to_tables.R
\name{ideal_island_to_tables}
\alias{ideal_island_to_tables}
\title{Convert an \code{ideal_island} to a list of tables}
\usage{
ideal_island_to_tables(ideal_island)
}
\arguments{
\item{ideal_island}{a simulated island with perfect information,
which is a \link{list} of one or more \code{ideal_island_clade}s.

An \code{ideal_island} is created by \link{sim_island} (together with \code{empirical_island},
which has imperfect information),
as can be checked by \link{check_ideal_island}}
}
\value{
a \link{list} with elements:
\itemize{
\item \code{speciations}: a table with the speciation events
\item \code{colonisations}: a table with colonisations
}
}
\description{
Convert an \code{ideal_island} to a list of tables
}
