% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/island_to_tables.R
\name{island_to_tables}
\alias{island_to_tables}
\title{Convert an \code{island} to a list of tables}
\usage{
island_to_tables(island)
}
\value{
a \link{list} with elements
}
\description{
Convert an \code{island} to a list of tables
}
