% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctt.R
\name{plot_ctt}
\alias{plot_ctt}
\title{Plots a heatmap of the colonisation through time statistic for mainland
sampling probability (y-axis) and mainland extinction rate (x-axis).}
\usage{
plot_ctt()
}
\value{
Void (saves plot)
}
\description{
Plots a heatmap of the colonisation through time statistic for mainland
sampling probability (y-axis) and mainland extinction rate (x-axis).
}
