% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctt_heatmap.R
\name{plot_ctt_heatmap}
\alias{plot_ctt_heatmap}
\title{Plots a heatmap of the colonisation through time statistic for mainland
sampling probability (y-axis) and mainland extinction rate (x-axis).}
\usage{
plot_ctt_heatmap(analysis_results, output_file_path)
}
\arguments{
\item{analysis_results}{A list of \code{analysis_result} elements which can be
checked with \code{check_analysis_result}}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}
}
\value{
Void (saves plot)
}
\description{
Plots a heatmap of the colonisation through time statistic for mainland
sampling probability (y-axis) and mainland extinction rate (x-axis).
}
