% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ctt_scatter.R
\name{plot_ctt_scatter}
\alias{plot_ctt_scatter}
\title{Plots a heatmap of the colonisation through time statistic for mainland
sampling probability (y-axis) and mainland extinction rate (x-axis).}
\usage{
plot_ctt_scatter(data_folder_path, output_file_path, parameter)
}
\arguments{
\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{parameter}{String specifying which parameter is plotted.
"mainland_ex" plots mainland extinction as the variable,
"mainland_sample_prob" plots mainland sampling probability as the variable
and "both" plots all the data}
}
\value{
Void (saves plot)
}
\description{
Plots a heatmap of the colonisation through time statistic for mainland
sampling probability (y-axis) and mainland extinction rate (x-axis).
}
