% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_daisie_data.R
\name{plot_daisie_data}
\alias{plot_daisie_data}
\title{Plot the DAISIE data}
\usage{
plot_daisie_data(daisie_data)
}
\arguments{
\item{daisie_data}{List containing data of DAISIE simulation with mainland
dynamics. Output from \code{sim_island_with_mainland}.

A list. The highest level of the list has two elements called
\code{ideal_islands} and \code{empirical_islands} which corresponds to
the ideal and empirical data sets produce in each simulation. Within each
\code{ideal_islands} and \code{empirical_islands} each element is an
individual replicate. The first element of each replicate is composed of
island information containing:
\describe{
\item{\verb{$island_age}}{A numeric with the island age.}
\item{\verb{$not_present}}{the number of mainland lineages that are not
present on the island.}
}
The subsequent elements of the list pertaining to each replicate contain
information on a single colonist lineage on the island and have 3 components:
\describe{
\item{\verb{$branching_times}}{island age and stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge and
Endemic anagenetic species.
For cladogenetic species these should
be island age and branching times of the radiation including the
stem age of the radiation.}
\item{\verb{$stac}}{An integer ranging from 1 to 6
indicating the status of the colonist:
\enumerate{
\item Non_endemic_MaxAge
\item Endemic
\item Endemic&Non_Endemic
\item Non_endemic_MaxAge
\item Endemic_singleton_MaxAge
\item Endemic_clade_MaxAge
}}
\item{\code{missing_species}}{number of island species that were
not sampled for particular clade (only applicable for endemic clades)}
}}
}
\value{
a \code{ggplot2}
}
\description{
Plot the DAISIE data
}
\examples{
set.seed(
  9,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
total_time <- 1
m <- 10 # number of mainland clades
mainland <- sim_mainland(
  total_time = total_time,
  m = m,
  mainland_ex = 2.0
)
mainland_clade <- mainland[[1]]
plot_mainland_clade(mainland_clade)
island <- DAISIEmainland:::sim_island(
  total_time = total_time,
  island_pars = c(1, 1, 10, 12, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
plot_island(island)
daisie_data <- DAISIEmainland:::format_to_daisie_data(
  island_replicates = island,
  total_time = total_time,
  m = m
)
# Not yet plot_daisie_data(daisie_data)
}
\seealso{
These are the functions to plot an evolutionary history:
\itemize{
\item Use \link{plot_mainland} to plot the mainland history
\item Use \link{plot_island} to plot the island history.
based on both empirical data and ideal data.
\item Use \link{plot_empirical_island} to plot the island history.
based on empirical data.
\item Use \link{plot_ideal_island} to plot the island history
based on ideal data.
}
}
\author{
Richèl J.C. Bilderbeek
}
