% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_empirical_island.R
\name{plot_empirical_island}
\alias{plot_empirical_island}
\title{Plot the simulated ideal island.}
\usage{
plot_empirical_island(empirical_island)
}
\arguments{
\item{empirical_island}{a simulated island with imperfect information,
as created by \link{sim_island} (together with \code{ideal_island},
which has perfect information).

\code{empirical_island} is a list  containing 3 components:
\itemize{
\item \verb{$branching_times}: island age and stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge and
Endemic anagenetic species.

For cladogenetic species these should
be island age and branching times of the radiation including the
stem age of the radiation.
\item \verb{$stac}: An integer ranging from 1 to 6
indicating the status of the colonist:
\enumerate{
\item Non_endemic_MaxAge
\item Endemic
\item Endemic&Non_Endemic
\item Non_endemic
\item Endemic_singleton_MaxAge
\item Endemic_clade_MaxAge
}
\item \verb{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)
}}
}
\value{
a \code{ggplot2}
}
\description{
Plot the simulated ideal island.
}
\examples{
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = 2)
island <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 1, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
plot_empirical_island(island$empirical_island)
}
\author{
Richèl J.C. Bilderbeek
}
