% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_endemics.R
\name{plot_endemics}
\alias{plot_endemics}
\title{Plots two faceted violin plots of the percent of endemics on the island for
different values of mainland extinction. The left side facet has the ideal
data and the right side facet has the empirical data.}
\usage{
plot_endemics(test = FALSE)
}
\arguments{
\item{test}{Boolean determining whether a plot is being tested}
}
\value{
Plot
}
\description{
Plots two faceted violin plots of the percent of endemics on the island for
different values of mainland extinction. The left side facet has the ideal
data and the right side facet has the empirical data.
}
