% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_endemics.R
\name{plot_endemics}
\alias{plot_endemics}
\title{Plots two faceted boxplots of the percent of endemics on the island for
different values of mainland extinction. The left side facet has the ideal
data and the right side facet has the empirical data.}
\usage{
plot_endemics(analysis_results, output_file_path, parameter, labels = NULL)
}
\arguments{
\item{analysis_results}{A list of analysis result elements which can be
checked with \code{check_analysis_result}}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{parameter}{String specifying which parameter is plotted.
"mainland_ex" plots mainland extinction as the variable,
"mainland_sample_prob" plots mainland sampling probability as the variable
and "both" plots all the data}

\item{labels}{A string or vector of strings to label the plotting grid}
}
\value{
Plot
}
\description{
Plots two faceted boxplots of the percent of endemics on the island for
different values of mainland extinction. The left side facet has the ideal
data and the right side facet has the empirical data.
}
