% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ideal_island.R
\name{plot_ideal_island}
\alias{plot_ideal_island}
\title{Plot the simulated ideal island.}
\usage{
plot_ideal_island(ideal_island)
}
\arguments{
\item{ideal_island}{a simulated island with perfect information,
as created by \link{sim_island} (together with \code{empirical_island},
which has imperfect information).

\code{ideal_island} is a list  containing 3 components:
\itemize{
\item \verb{$branching_times}: island age and stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge and
Endemic anagenetic species.

For cladogenetic species these should
be island age and branching times of the radiation including the
stem age of the radiation.
\item \verb{$stac}: An integer ranging from 1 to 6
indicating the status of the colonist:
\enumerate{
\item Non_endemic_MaxAge
\item Endemic
\item Endemic&Non_Endemic
\item Non_endemic
\item Endemic_singleton_MaxAge
\item Endemic_clade_MaxAge
}
\item \verb{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)
}

For recolonising lineages, there is an extra element,
\code{all_colonisations} per list element.
It is comprised of \verb{$event_times} and \verb{$species_type}:
\describe{
\item{\verb{$event_times}}{ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times. Each vector pertains to one
colonising lineage.}
\item{\verb{$species_type}}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}}
}
\value{
a \code{ggplot2}
}
\description{
Plot the simulated ideal island.
}
\examples{
mainland_clade <- DAISIEmainland:::create_test_mainland_clade(
  mainland_scenario = 2)
island <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 1, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
plot_ideal_island(island$ideal_island)

}
\seealso{
These are the functions to plot an evolutionary history:
\itemize{
\item Use \link{plot_mainland} to plot the mainland history
\item Use \link{plot_island} to plot the island history.
based on both empirical data and ideal data.
\item Use \link{plot_empirical_island} to plot the island history.
based on empirical data.
\item Use \link{plot_ideal_island} to plot the island history
based on ideal data.
}
}
\author{
Richèl J.C. Bilderbeek
}
