% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ideal_island.R
\name{plot_ideal_island}
\alias{plot_ideal_island}
\title{Plot the simulated mainland.}
\usage{
plot_ideal_island(ideal_island)
}
\value{
a \code{ggplot2}
}
\description{
Plot the simulated mainland.
}
\examples{
mainland_clade <- create_test_mainland_clade(mainland_scenario = 2)
island <- sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 1, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
plot_ideal_island(island$ideal_island)
}
\author{
Richèl J.C. Bilderbeek
}
