% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inf_k.R
\name{plot_inf_k}
\alias{plot_inf_k}
\title{Plots the percentage of carrying capacity (K') estimates that are infinite}
\usage{
plot_inf_k(analysis_results, output_file_path, labels = NULL)
}
\arguments{
\item{analysis_results}{A list of analysis result elements which can be
checked with \code{check_analysis_result}}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{labels}{A string or vector of strings to label the plotting grid}
}
\value{
Void (saves plot)
}
\description{
Plots the percentage of carrying capacity (K') estimates that are infinite
}
