% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inf_k.R
\name{plot_inf_k}
\alias{plot_inf_k}
\title{Plots the percentage of carrying capacity (K') estimates that are infinite}
\usage{
plot_inf_k(data_folder_path, output_file_path, param_space, parameter)
}
\arguments{
\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{param_space}{String specifying which parameter space results to read.
Either "general", "mainland_ex", or "mainland_sample_prob"}

\item{parameter}{String specifying which parameter is plotted.
"mainland_ex" plots mainland extinction as the variable,
"mainland_sample_prob" plots mainland sampling probability as the variable
and "both" plots all the data}
}
\value{
Void (saves plot)
}
\description{
Plots the percentage of carrying capacity (K') estimates that are infinite
}
