% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_island.R
\name{plot_island}
\alias{plot_island}
\title{Plot the simulated island, based on ideal and empirical data.}
\usage{
plot_island(island)
}
\arguments{
\item{island}{The history of a a single island.
A \link{list} with two elements:
\itemize{
\item \code{ideal_island}: the island history based on ideal data
\item \code{empirical_island}: the island history based on empirical data
}

Each of these is a \link{list} with the following elements:
\itemize{
\item \verb{$branching_times}: island age and stem age of the
population/species in the case of Non-endemic, Non-endemic_MaxAge and
Endemic anagenetic species.

For cladogenetic species these should
be island age and branching times of the radiation including the
stem age of the radiation.
\item \verb{$stac}: An integer ranging from 1 to 6
indicating the status of the colonist:
\enumerate{
\item Non_endemic_MaxAge
\item Endemic
\item Endemic&Non_Endemic
\item Non_endemic
\item Endemic_singleton_MaxAge
\item Endemic_clade_MaxAge
}
\item \verb{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)
}

For recolonising lineages, there is an extra element,
\code{all_colonisations} per list element.
It is comprised of \verb{$event_times} and \verb{$species_type}:
\describe{
\item{\verb{$event_times}}{ordered numeric vectors containing all
events for each extant recolonising lineage. This includes all
colonisation and branching times. Each vector pertains to one
colonising lineage.}
\item{\verb{$species_type}}{a string. Can be \code{"A"}, \code{"C"} or
\code{"I"} depending on whether the extant clade is of anagenetic,
cladogenetic or immigrant origin, respectively.}
}}
}
\value{
a \code{ggplot2}
}
\description{
Plot the simulated island, based on ideal and empirical data.
}
\details{
The plot will consist of a grid,
with two columns and one or more rows.
The columns depict the same evolutionary history of the
island, yet for different type of data.
The rows depict the species clades.

The vertical lines in the column for the ideal data depict
one or more colonisation events.
}
\examples{
set.seed(
  9,
  kind = "Mersenne-Twister",
  normal.kind = "Inversion",
  sample.kind = "Rejection"
)
mainland <- sim_mainland(
  total_time = 10,
  m = 10,
  mainland_ex = 1.0
)
mainland_clade <- mainland[[1]]
plot_mainland_clade(mainland_clade)
island <- DAISIEmainland:::sim_island(
  total_time = 1,
  island_pars = c(1, 1, 10, 12, 1),
  mainland = mainland_clade,
  mainland_sample_prob = 1,
  mainland_sample_type = "complete")
plot_island(island)

}
\seealso{
These are the functions to plot an evolutionary history:
\itemize{
\item Use \link{plot_mainland} to plot the mainland history
\item Use \link{plot_island} to plot the island history.
based on both empirical data and ideal data.
\item Use \link{plot_empirical_island} to plot the island history.
based on empirical data.
\item Use \link{plot_ideal_island} to plot the island history
based on ideal data.
}
}
\author{
Richèl J.C. Bilderbeek
}
