% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_k_estimates.R
\name{plot_k_estimates}
\alias{plot_k_estimates}
\title{Plots carrying capacity estimates}
\usage{
plot_k_estimates(data_folder_path, output_file_path, parameter)
}
\arguments{
\item{data_folder_path}{String specifying the directory the data is read
from}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{parameter}{String specifying which parameter is plotted.
"mainland_ex" plots mainland extinction as the variable,
"mainland_sample_prob" plots mainland sampling probability as the variable
and "both" plots all the data}
}
\value{
Void (saves plot)
}
\description{
Plots carrying capacity estimates
}
