% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_k_estimates.R
\name{plot_k_estimates}
\alias{plot_k_estimates}
\title{Boxplots of the estimated carrying capacity faceted for ideal and empirical
data for two different values of K'.}
\usage{
plot_k_estimates(
  analysis_results,
  output_file_path,
  parameter,
  num_breaks,
  signif,
  scientific,
  labels = NULL
)
}
\arguments{
\item{analysis_results}{A list of analysis result elements which can be
checked with \code{check_analysis_result}}

\item{output_file_path}{String specifying the directory the file is saved
in, if NULL the file is returned to console and not saved}

\item{parameter}{String specifying which parameter is plotted.
"mainland_ex" plots mainland extinction as the variable,
"mainland_sample_prob" plots mainland sampling probability as the variable
and "both" plots all the data}

\item{num_breaks}{A numeric specifying how many breaks are wanted on the
plot axes}

\item{signif}{A numeric specifying how many significant figures the axes
labels have when plotting}

\item{scientific}{A boolean determining whether the axis labels will be
converted to scientific notation}

\item{labels}{A string or vector of strings to label the plotting grid}
}
\value{
Void (saves plot)
}
\description{
Boxplots of the estimated carrying capacity faceted for ideal and empirical
data for two different values of K'.
}
