% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mainland.R
\name{plot_mainland}
\alias{plot_mainland}
\title{Plot the simulated mainland.}
\usage{
plot_mainland(mainland)
}
\arguments{
\item{mainland}{the evolutionary history of the mainland species,
as created by \link{sim_mainland}.
Use \link{plot_mainland} to visualise the that evolutionary history.}
}
\value{
a \code{ggplot2} facet plot, where each facet/row contains
the evolutionary history of a clade.
Per clade, the x axis shows the time from past (at the left)
to the present (at the right).
The y-axis has no meaning except for separating the different
species using \link{branch_code_to_y}.
}
\description{
Plot the simulated mainland.
}
\examples{
mainland <- DAISIEmainland::sim_mainland(
  total_time = 1,
  m = 2,
  mainland_ex = 2
)

plot_mainland(mainland)
}
\seealso{
These are the functions to plot an evolutionary history:
\itemize{
\item Use \link{plot_mainland} to plot the mainland history
}
}
\author{
Richèl J.C. Bilderbeek
}
