% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_max_age.R
\name{plot_max_age}
\alias{plot_max_age}
\title{Plots two faceted violin plots of the percent of max age colonists on the
island for different values of mainland extinction. The left side facet has
the ideal data and the right side facet has the empirical data.}
\usage{
plot_max_age(test = FALSE)
}
\arguments{
\item{test}{Boolean determining whether a function is being tested}
}
\value{
Plot
}
\description{
Plots two faceted violin plots of the percent of max age colonists on the
island for different values of mainland extinction. The left side facet has
the ideal data and the right side facet has the empirical data.
}
