% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_param_estimates.R
\name{plot_param_estimates}
\alias{plot_param_estimates}
\title{Plots a grid of density plots of parameter estimates on the diagonal and
scatter plots of parameter estimates under the diagonal.}
\usage{
plot_param_estimates(param_set, xlim = FALSE)
}
\arguments{
\item{param_set}{Numeric for the index of the parameter set from the
parameter space}

\item{xlim}{Boolean determining whether the x-axis should be bounded on the
plot}
}
\value{
Void (saves plot)
}
\description{
Plots a grid of density plots of parameter estimates on the diagonal and
scatter plots of parameter estimates under the diagonal.
}
