% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mainland.R
\name{sample_mainland}
\alias{sample_mainland}
\title{Samples the mainland clade with the mainland sampling probability}
\usage{
sample_mainland(total_time, mainland_clade, mainland_sample_prob, island_spec)
}
\arguments{
\item{total_time}{Numeric defining the length of the simulation in time units.}

\item{mainland_clade}{Data frame with state of mainland. see
\code{\link{create_test_mainland_clade}}.}

\item{mainland_sample_prob}{Numeric between zero and one determining the
probability of an extant mainland species being sampled.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}
}
\value{
Matrix
}
\description{
Samples the mainland clade with the mainland sampling probability
}
