% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mainland.R
\name{sample_mainland}
\alias{sample_mainland}
\title{Samples the mainland clade with the mainland sampling probability}
\usage{
sample_mainland(totaltime, mainland, mainland_sample_prob, island_spec)
}
\arguments{
\item{totaltime}{Numeric defining the length of the simulation in time units.}

\item{mainland}{Matrix with state of mainland.}

\item{mainland_sample_prob}{Numeric between zero and one determining the
probability of an extant mainland species being sampled.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}
}
\value{
Matrix
}
\description{
Samples the mainland clade with the mainland sampling probability
}
