% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{scientific}
\alias{scientific}
\title{Creates the axis numbers (labels) for plotting with scientific form in
x10 notation}
\usage{
scientific(breaks, scientific, signif)
}
\arguments{
\item{breaks}{A vector of numerics}

\item{scientific}{A boolean determining whether the axis labels will be
converted to scientific notation}

\item{signif}{A numeric specifying how many significant figures the axes
labels have when plotting}
}
\value{
Character vector
}
\description{
Creates the axis numbers (labels) for plotting with scientific form in
x10 notation
}
