% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_island.R
\name{sim_island}
\alias{sim_island}
\title{Simulates the island.}
\usage{
sim_island(time, m, island_pars, mainland_clade, mainland_sample_prob)
}
\arguments{
\item{time}{Numeric defining the length of the simulation in time units.}

\item{m}{Numeric defining the size of mainland pool.}

\item{island_pars}{A numeric vector containing the parameters for the island:
\itemize{
  \item{\code{island_pars[1]}: lambda^c (cladogenesis rate)}
  \item{\code{island_pars[2]}: mu (extinction rate)}
  \item{\code{island_pars[3]}: K (carrying capacity), set K=Inf for
  diversity independence.}
  \item{\code{island_pars[4]}: gamma (immigration rate)}
  \item{\code{island_pars[5]}: lambda^a (anagenesis rate)}
  }}

\item{mainland_clade}{Data frame with state of mainland. see
\code{\link{create_test_mainland_clade}}.}

\item{mainland_sample_prob}{Numeric between zero and one determining the
probability of an extant mainland species being sampled.}
}
\value{
List
}
\description{
Simulates the island.
}
\keyword{internal}
