% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac_to_str.R
\name{stac_to_str}
\alias{stac_to_str}
\title{Convert a \code{stac} (status of colonist) to a string}
\usage{
stac_to_str(stac)
}
\value{
the event as a string
}
\description{
Convert a \code{stac} (status of colonist) to a string
}
\examples{
DAISIEmainland:::stac_to_str(1)
DAISIEmainland:::stac_to_str(2)
DAISIEmainland:::stac_to_str(3)
DAISIEmainland:::stac_to_str(4)
DAISIEmainland:::stac_to_str(5)
DAISIEmainland:::stac_to_str(6)
}
\seealso{
use \link{str_to_event} to convent a string to an \code{event}
}
\author{
Richèl J.C. Bilderbeek
}
\keyword{internal}
