% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_island_endemics.R
\name{update_island_endemics}
\alias{update_island_endemics}
\title{Checks and updates if any changes on the mainland change the state of the
island species.}
\usage{
update_island_endemics(timeval, total_time, island_spec, mainland_clade)
}
\arguments{
\item{timeval}{Numeric defining current time of simulation.}

\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{island_spec}{Matrix with current state of simulation containing number
of species, see \code{\link[=create_test_island_spec]{create_test_island_spec()}}}

\item{mainland_clade}{Data frame with state of mainland, see
\code{\link[=create_test_mainland_clade]{create_test_mainland_clade()}}}
}
\value{
Data frame of island species
}
\description{
Checks and updates if any changes on the mainland change the state of the
island species.
}
\author{
Joshua W. Lambert
}
