% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_island_endemics.R
\name{update_island_endemics}
\alias{update_island_endemics}
\title{Checks and updates if any changes on the mainland change the state of the
island species.}
\usage{
update_island_endemics(timeval, totaltime, island_spec, mainland)
}
\arguments{
\item{timeval}{Numeric defining current time of simulation.}

\item{totaltime}{Numeric defining the length of the simulation in time units.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}

\item{mainland}{Data frame with state of mainland. see
\code{\link{create_test_mainland}}.}
}
\value{
Matrix or NULL
}
\description{
Checks and updates if any changes on the mainland change the state of the
island species.
}
